refactor(hub): delete the embedded MCP server
Issue 763 P5e. The MCP harness moved out-of-process into the
aether-mcp crate (P5b/c) and .mcp.json pointed at it in P5d, leaving
the embedded hub MCP server (port 8888) as dead weight. P5e deletes it.

- Delete hub/mcp.rs + hub/mcp/{args,codecs,tools,trace}.rs (~5k LOC).
- Surgically remove MCP wiring from hub/chassis.rs: mcp_addr drops
  from HubEnv + HubServerDriverCapability + HubServerDriverRunning,
  the coordinator stops spawning run_mcp_server + the matching
  select! arm, HubState construction goes. The legacy engine TCP
  listener + the loopback drainers + ProcessCapability stay — those
  retire with the chassis collapse in P5f.
- Strip rmcp / axum / base64 / schemars from
  aether-substrate-bundle/Cargo.toml (they had zero remaining real
  usage outside hub/mcp/).
- Relocate ComponentCapabilitiesWire / HandlerCapabilityWire /
  FallbackCapabilityWire from hub/mcp/args.rs into hub/registry.rs
  (their only consumer), dropping the JsonSchema derive that pulled
  schemars in.
- CLAUDE.md "## MCP harness" rewritten: aether-mcp is the only path
  now; tool bullets rewritten to match aether-mcp's actual surface
  (9 of the prior 11).

Verified: cargo nextest run --workspace --retries 2 — 1084 passed
(49 fewer than baseline; those were unit tests inside the deleted
hub/mcp/ module). clippy / fmt clean.

Part of iamacoffeepot/aether#763 (P5e). Next: P5f — collapse the old
engine-listener model (engine.rs, wire.rs, session.rs, loopback
drainers, ProcessCapability), rewrite the chassis driver, delete the
dead engine_handshake/hub_client/engine_logs tests, file the two
out-of-scope follow-ups (log pull migration, broadcast retirement),
and close iamacoffeepot/aether#763.
