+++ crates/aether-substrate-bundle/Cargo.toml
@@ -45,19 +45,14 @@ aether-kinds = { path = "../aether-kinds" }
 # bridge BootError / wasmtime / io faults purely to bubble up.
 # Same crate as `wasmtime::Result` re-exports (issue #571).
 anyhow = "1"
-axum = "0.8"
-base64 = "0.22"
 bytemuck = { version = "1.19", features = ["derive"] }
 os_info = { version = "3", default-features = false }
 png = "0.17"
 pollster = "0.4.0"
 postcard = { version = "1.1", default-features = false, features = ["alloc"] }
-rmcp = { version = "0.8", features = [
-    "server",
-    "macros",
-    "transport-streamable-http-server",
-] }
-schemars = "1"
+# `rmcp` / `axum` / `base64` / `schemars` retired with the embedded MCP
+# server (issue 763 P5e) — the harness moved out-of-process into the
+# `aether-mcp` crate, which owns those deps now.
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "net", "io-util", "sync", "time", "signal", "process", "fs"] }
