--- crates/aether-substrate-bundle/src/hub/mcp.rs
@@ -1,2088 +1,0 @@
-//! Claude-facing MCP tool surface (ADR-0006 V0, ADR-0007, ADR-0008,
-//! ADR-0009). Exposes eight tools for sending and receiving mail,
-//! introspecting engine state, spawning and terminating substrates,
-//! draining logs, and capturing frames.
-//!
-//! The rmcp `Service` factory is invoked per session, so `Hub` is cheap
-//! to clone and shares a single `HubState` via `Arc`. Per-tool output
-//! is returned as a JSON-encoded `String`; rmcp wraps it into a
-//! `Content::text` automatically via `IntoContents`.
-//!
-//! Submodule split:
-//!   - `args`: request/response shapes (pure data).
-//!   - `codecs`: schema-driven encode/decode between tool JSON and
-//!     wire bytes.
-//!   - `tools`: the `#[tool_router] impl Hub` block and the
-//!     `ServerHandler` impl.
-
-use std::net::SocketAddr;
-use std::sync::Arc;
-
-use rmcp::{
-    handler::server::router::tool::ToolRouter,
-    transport::streamable_http_server::{
-        StreamableHttpServerConfig, StreamableHttpService, session::local::LocalSessionManager,
-    },
-};
-use tokio::sync::{Mutex, mpsc};
-
-use crate::hub::log_store::LogStore;
-use crate::hub::registry::EngineRegistry;
-use crate::hub::session::{QueuedMail, SessionHandle, SessionRegistry};
-
-pub(crate) mod args;
-mod codecs;
-mod tools;
-mod trace;
-
-// Bring args types and commonly-referenced protocol types into mcp's
-// namespace so the tests module below can reach them via `use
-// super::*;` without repeating imports for every case. Internal
-// helpers in `codecs` stay `pub(super)` and are reachable from
-// `tools` only.
-#[cfg(test)]
-use crate::hub::wire::{HubToEngine, SessionToken, Uuid};
-#[cfg(test)]
-use args::{
-    CaptureFrameArgs, DescribeKindsArgs, EngineInfo, MailSpec, MailStatus, ReceiveMailArgs,
-    ReceivedMail, SendMailArgs,
-};
-#[cfg(test)]
-use base64::Engine as _;
-#[cfg(test)]
-use rmcp::handler::server::wrapper::Parameters;
-#[cfg(test)]
-use std::collections::HashMap;
-
-/// Default port the hub binds for MCP clients. Overridable via
-/// `AETHER_MCP_PORT`.
-pub const DEFAULT_MCP_PORT: u16 = 8888;
-
-/// Shared state across all rmcp sessions. Cheap to `Arc::clone` into
-/// each per-session `Hub` instance.
-///
-/// Post-ADR-0078 Phase 1 (PR 596): `pending_spawns` and `hub_engine_addr`
-/// retired — process supervision moved into [`crate::hub::process_capability::ProcessCapability`],
-/// which is configured with both at chassis-builder time. The hub's
-/// engine handshake path still references the same `PendingSpawns`
-/// instance the cap owns, but the MCP tool surface no longer needs
-/// either field.
-pub struct HubState {
-    pub(crate) engines: EngineRegistry,
-    pub(crate) sessions: SessionRegistry,
-    /// ADR-0023 per-engine log buffers. Outlives engine records so
-    /// post-mortem `engine_logs` polls succeed after a substrate exit.
-    pub(crate) logs: LogStore,
-}
-
-impl HubState {
-    pub fn new(engines: EngineRegistry, sessions: SessionRegistry, logs: LogStore) -> Arc<Self> {
-        Arc::new(Self {
-            engines,
-            sessions,
-            logs,
-        })
-    }
-}
-
-/// Per-session rmcp service. rmcp calls the factory once per MCP
-/// session and may clone the result for concurrent tool dispatch;
-/// `session` is an `Arc<SessionHandle>` so the registry entry only
-/// goes away when the last clone drops.
-#[derive(Clone)]
-pub struct Hub {
-    pub(crate) state: Arc<HubState>,
-    pub(crate) tool_router: ToolRouter<Self>,
-    pub(crate) session: Arc<SessionHandle>,
-    /// Drain for this session's inbound observation mail.
-    /// `receive_mail` pulls from it non-blocking; wrapping in an
-    /// `Arc<Mutex<_>>` lets rmcp's per-tool-call clones share the same
-    /// receiver.
-    pub(crate) inbound: Arc<Mutex<mpsc::Receiver<QueuedMail>>>,
-}
-
-// `Hub::new` lives in `tools.rs` next to the `#[tool_router]` impl —
-// rmcp's macro emits a private `Self::tool_router()` that the
-// constructor calls, and a sibling impl block in the same module can
-// reach it without widening the macro-generated visibility.
-
-/// Bind an axum server on `addr` exposing the MCP tool surface at
-/// `/mcp`. Returns on axum error. The caller owns the cancellation.
-pub async fn run_mcp_server(addr: SocketAddr, state: Arc<HubState>) -> std::io::Result<()> {
-    let factory_state = Arc::clone(&state);
-    let service = StreamableHttpService::new(
-        move || Ok(Hub::new(Arc::clone(&factory_state))),
-        Arc::new(LocalSessionManager::default()),
-        StreamableHttpServerConfig::default(),
-    );
-
-    let app = axum::Router::new().nest_service("/mcp", service);
-    let listener = tokio::net::TcpListener::bind(addr).await?;
-    let bound = listener.local_addr()?;
-    eprintln!("aether-substrate-hub: mcp listener bound on http://{bound}/mcp");
-    axum::serve(listener, app).await
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::hub::registry::EngineRecord;
-    use crate::hub::wire::EngineId;
-    use tokio::sync::mpsc;
-
-    /// Build a `HubState` for tests that don't exercise process
-    /// supervision. Post-ADR-0078 Phase 1, MCP tools route the
-    /// spawn / terminate path through `ProcessCapability`; tests
-    /// targeting that path do so via the cap's own unit tests in
-    /// `process_capability.rs`.
-    fn test_state(engines: EngineRegistry, sessions: SessionRegistry) -> Arc<HubState> {
-        HubState::new(engines, sessions, LogStore::new())
-    }
-
-    fn record(id_u128: u128) -> (EngineRecord, mpsc::Receiver<HubToEngine>) {
-        // Default kinds: just `aether.tick` as Schema(Unit) so tests
-        // that don't care about a specific schema can use the default
-        // record and still send tick mail. ADR-0019 PR 5 removed
-        // `payload_bytes`, so a kind without a descriptor is now
-        // unreachable from `send_mail`.
-        let tick = aether_data::KindDescriptor {
-            name: "aether.tick".into(),
-            schema: aether_data::SchemaType::Unit,
-        };
-        record_with_kinds(id_u128, vec![tick])
-    }
-
-    fn record_with_kinds(
-        id_u128: u128,
-        kinds: Vec<aether_data::KindDescriptor>,
-    ) -> (EngineRecord, mpsc::Receiver<HubToEngine>) {
-        let (tx, rx) = mpsc::channel(16);
-        let rec = EngineRecord {
-            id: EngineId(Uuid::from_u128(id_u128)),
-            name: format!("engine-{id_u128}"),
-            pid: 42,
-            version: "test".into(),
-            kinds,
-            mailboxes: vec![],
-            components: HashMap::new(),
-            mail_tx: tx,
-            spawned: false,
-        };
-        (rec, rx)
-    }
-
-    #[tokio::test]
-    async fn list_engines_reflects_registry() {
-        let engines = EngineRegistry::new();
-        let (a, _rx_a) = record(1);
-        let (b, _rx_b) = record(2);
-        engines.insert(a);
-        engines.insert(b);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(state);
-
-        let json = hub.list_engines().await.unwrap();
-        let list: Vec<EngineInfo> = serde_json::from_str(&json).unwrap();
-        assert_eq!(list.len(), 2);
-    }
-
-    fn spec(engine_id: String, kind: &str, params: Option<serde_json::Value>) -> MailSpec {
-        MailSpec {
-            engine_id,
-            recipient_name: "hello".into(),
-            kind_name: kind.into(),
-            params,
-            count: 1,
-        }
-    }
-
-    async fn run(hub: &Hub, mails: Vec<MailSpec>) -> Vec<MailStatus> {
-        let json = hub
-            .send_mail(Parameters(SendMailArgs { mails }))
-            .await
-            .unwrap();
-        serde_json::from_str(&json).unwrap()
-    }
-
-    #[tokio::test]
-    async fn send_mail_stamps_sender_with_session_token() {
-        let engines = EngineRegistry::new();
-        let sessions = SessionRegistry::new();
-        let (rec, mut rx) = record(100);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, sessions));
-        let expected = hub.session.token;
-
-        let statuses = run(&hub, vec![spec(id.0.to_string(), "aether.tick", None)]).await;
-        assert_eq!(statuses[0].status, "delivered");
-
-        let HubToEngine::Mail(m) = rx.try_recv().unwrap() else {
-            panic!()
-        };
-        assert_eq!(m.sender, expected);
-        assert_ne!(m.sender, SessionToken::NIL);
-    }
-
-    #[tokio::test]
-    async fn two_hubs_mint_distinct_session_tokens() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let a = Hub::new(Arc::clone(&state));
-        let b = Hub::new(Arc::clone(&state));
-        assert_ne!(a.session.token, b.session.token);
-        // Both live in the registry.
-        assert_eq!(state.sessions.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn dropping_hub_deregisters_session() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let token = hub.session.token;
-        assert!(state.sessions.get(&token).is_some());
-        drop(hub);
-        assert!(state.sessions.get(&token).is_none());
-    }
-
-    #[tokio::test]
-    async fn send_mail_params_encodes_via_descriptor() {
-        use aether_data::{KindDescriptor, NamedField, Primitive, SchemaType};
-        let engines = EngineRegistry::new();
-        let kinds = vec![KindDescriptor {
-            name: "aether.mouse_move".into(),
-            schema: SchemaType::Struct {
-                repr_c: true,
-                fields: vec![
-                    NamedField {
-                        name: "x".into(),
-                        ty: SchemaType::Scalar(Primitive::F32),
-                    },
-                    NamedField {
-                        name: "y".into(),
-                        ty: SchemaType::Scalar(Primitive::F32),
-                    },
-                ]
-                .into(),
-            },
-        }];
-        let (rec, mut rx) = record_with_kinds(3, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-
-        let statuses = run(
-            &hub,
-            vec![spec(
-                id.0.to_string(),
-                "aether.mouse_move",
-                Some(serde_json::json!({"x": 10.5, "y": 20.0})),
-            )],
-        )
-        .await;
-        assert_eq!(statuses[0].status, "delivered");
-
-        let HubToEngine::Mail(m) = rx.try_recv().unwrap() else {
-            panic!()
-        };
-        let mut expected = Vec::new();
-        expected.extend_from_slice(&10.5f32.to_le_bytes());
-        expected.extend_from_slice(&20.0f32.to_le_bytes());
-        assert_eq!(m.payload, expected);
-    }
-
-    #[tokio::test]
-    async fn send_mail_unit_kind_no_params() {
-        use aether_data::{KindDescriptor, SchemaType};
-        let engines = EngineRegistry::new();
-        let kinds = vec![KindDescriptor {
-            name: "aether.tick".into(),
-            schema: SchemaType::Unit,
-        }];
-        let (rec, mut rx) = record_with_kinds(4, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-
-        let statuses = run(&hub, vec![spec(id.0.to_string(), "aether.tick", None)]).await;
-        assert_eq!(statuses[0].status, "delivered");
-
-        let HubToEngine::Mail(m) = rx.try_recv().unwrap() else {
-            panic!()
-        };
-        assert!(m.payload.is_empty());
-    }
-
-    #[tokio::test]
-    async fn send_mail_batch_reports_per_item_status() {
-        let engines = EngineRegistry::new();
-        let (rec, mut rx) = record(5);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-
-        let good = spec(id.0.to_string(), "aether.tick", None);
-        let bad = spec(Uuid::from_u128(0xdead).to_string(), "aether.tick", None);
-        let good2 = spec(id.0.to_string(), "aether.tick", None);
-
-        let statuses = run(&hub, vec![good, bad, good2]).await;
-        assert_eq!(statuses.len(), 3);
-        assert_eq!(statuses[0].status, "delivered");
-        assert!(statuses[1].status.starts_with("error: unknown engine_id"));
-        assert_eq!(statuses[2].status, "delivered");
-
-        // Two frames actually went through.
-        assert!(rx.try_recv().is_ok());
-        assert!(rx.try_recv().is_ok());
-        assert!(rx.try_recv().is_err());
-    }
-
-    #[tokio::test]
-    async fn send_mail_unknown_kind_errors() {
-        // ADR-0019 cleanup: there's no fallback path for kinds without
-        // a descriptor — an agent gets a clear "no descriptor" error.
-        let engines = EngineRegistry::new();
-        let (rec, _rx) = record_with_kinds(9, vec![]);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-
-        let statuses = run(
-            &hub,
-            vec![spec(
-                id.0.to_string(),
-                "hello.unknown",
-                Some(serde_json::json!({})),
-            )],
-        )
-        .await;
-        assert!(
-            statuses[0].status.contains("no descriptor"),
-            "got: {}",
-            statuses[0].status
-        );
-    }
-
-    #[tokio::test]
-    async fn describe_kinds_returns_descriptors() {
-        use aether_data::{KindDescriptor, NamedField, Primitive, SchemaType};
-        let kinds = vec![
-            KindDescriptor {
-                name: "aether.tick".into(),
-                schema: SchemaType::Unit,
-            },
-            KindDescriptor {
-                name: "hello.note".into(),
-                schema: SchemaType::Struct {
-                    repr_c: false,
-                    fields: vec![NamedField {
-                        name: "body".into(),
-                        ty: SchemaType::String,
-                    }]
-                    .into(),
-                },
-            },
-            KindDescriptor {
-                name: "hello.cast".into(),
-                schema: SchemaType::Struct {
-                    repr_c: true,
-                    fields: vec![NamedField {
-                        name: "n".into(),
-                        ty: SchemaType::Scalar(Primitive::U32),
-                    }]
-                    .into(),
-                },
-            },
-        ];
-        let engines = EngineRegistry::new();
-        let (rec, _rx) = record_with_kinds(11, kinds.clone());
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(state);
-
-        let args = DescribeKindsArgs {
-            engine_id: id.0.to_string(),
-        };
-        let json = hub.describe_kinds(Parameters(args)).await.unwrap();
-        let back: Vec<KindDescriptor> = serde_json::from_str(&json).unwrap();
-        assert_eq!(back, kinds);
-    }
-
-    #[tokio::test]
-    async fn describe_kinds_unknown_engine_errors() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(state);
-        let args = DescribeKindsArgs {
-            engine_id: Uuid::from_u128(1).to_string(),
-        };
-        let err = hub.describe_kinds(Parameters(args)).await.unwrap_err();
-        assert!(format!("{err:?}").contains("unknown engine_id"));
-    }
-
-    // ADR-0033: `describe_component` returns stored capabilities for
-    // `(engine, mailbox)`. `load_component` normally populates the
-    // record; tests seed it directly via `upsert_component` to isolate
-    // the lookup path from the full load flow.
-
-    fn stub_capabilities() -> args::ComponentCapabilitiesWire {
-        use aether_data::tagged_id::Tag;
-        use aether_data::with_tag;
-        args::ComponentCapabilitiesWire {
-            handlers: vec![
-                args::HandlerCapabilityWire {
-                    id: aether_data::KindId(with_tag(Tag::Kind, 42)),
-                    name: "aether.tick".into(),
-                    doc: Some("heartbeat".into()),
-                },
-                args::HandlerCapabilityWire {
-                    id: aether_data::KindId(with_tag(Tag::Kind, 0xff)),
-                    name: "aether.ping".into(),
-                    doc: None,
-                },
-            ],
-            fallback: None,
-            doc: Some("A canary component.".into()),
-        }
-    }
-
-    #[tokio::test]
-    async fn describe_component_returns_stored_capabilities() {
-        use crate::hub::registry::ComponentRecord;
-        use aether_data::tagged_id::{self, Tag};
-        use aether_data::with_tag;
-
-        let engines = EngineRegistry::new();
-        let (rec, _rx) = record(50);
-        let id = rec.id;
-        engines.insert(rec);
-        let capabilities = stub_capabilities();
-        // ADR-0064: registry stores raw u64 with tag bits set; the
-        // wire form is the tagged-string encoding of that u64.
-        let mailbox_id = with_tag(Tag::Mailbox, 7);
-        engines.upsert_component(
-            &id,
-            aether_data::MailboxId(mailbox_id),
-            ComponentRecord {
-                name: "hello".into(),
-                capabilities: capabilities.clone(),
-            },
-        );
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-
-        let json = hub
-            .describe_component(Parameters(args::DescribeComponentArgs {
-                engine_id: id.0.to_string(),
-                mailbox_id: tagged_id::encode(mailbox_id).unwrap(),
-            }))
-            .await
-            .unwrap();
-        let response: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert_eq!(response["name"], "hello");
-        assert_eq!(response["doc"], "A canary component.");
-        let receives = response["receives"].as_array().unwrap();
-        assert_eq!(receives.len(), 2);
-        assert_eq!(receives[0]["name"], "aether.tick");
-        assert_eq!(receives[0]["doc"], "heartbeat");
-        assert_eq!(receives[1]["name"], "aether.ping");
-        assert!(receives[1]["doc"].is_null());
-        assert!(response["fallback"].is_null());
-        // Issue 690: handler ids surface as ADR-0064 tagged strings,
-        // not raw u64 numbers — JSON's f64 rounds 64-bit ids past 2^53
-        // so the wire form must be `knd-XXXX-XXXX-XXXX`.
-        let id0 = receives[0]["id"].as_str().expect("id must be a string");
-        assert!(id0.starts_with("knd-"), "id should be tagged: {id0}");
-        let id1 = receives[1]["id"].as_str().expect("id must be a string");
-        assert!(id1.starts_with("knd-"), "id should be tagged: {id1}");
-    }
-
-    #[tokio::test]
-    async fn describe_component_unknown_mailbox_errors() {
-        use aether_data::tagged_id::{self, Tag};
-        use aether_data::with_tag;
-
-        let engines = EngineRegistry::new();
-        let (rec, _rx) = record(51);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-
-        let bogus = tagged_id::encode(with_tag(Tag::Mailbox, 999)).unwrap();
-        let err = hub
-            .describe_component(Parameters(args::DescribeComponentArgs {
-                engine_id: id.0.to_string(),
-                mailbox_id: bogus.clone(),
-            }))
-            .await
-            .unwrap_err();
-        // The error message echoes the original tagged-string form so
-        // the agent sees back what they passed in (not a re-encoded
-        // u64). Asserting on the prefix keeps the test robust to the
-        // base32 body's exact bytes.
-        let msg = format!("{err:?}");
-        assert!(msg.contains("no component at mailbox_id"));
-        assert!(msg.contains(&bogus));
-    }
-
-    #[tokio::test]
-    async fn describe_component_unknown_engine_errors() {
-        use aether_data::tagged_id::{self, Tag};
-        use aether_data::with_tag;
-
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(state);
-
-        let err = hub
-            .describe_component(Parameters(args::DescribeComponentArgs {
-                engine_id: Uuid::from_u128(0xdead).to_string(),
-                mailbox_id: tagged_id::encode(with_tag(Tag::Mailbox, 0)).unwrap(),
-            }))
-            .await
-            .unwrap_err();
-        assert!(format!("{err:?}").contains("unknown engine_id"));
-    }
-
-    /// Issue 718: descriptor for `aether.trace.describe_tree`.
-    /// Cloned from the kind's `Schema` const so a future schema change
-    /// flows through to the test for free.
-    fn describe_tree_descriptor() -> aether_data::KindDescriptor {
-        use aether_data::Kind as _;
-        aether_data::KindDescriptor {
-            name: aether_kinds::trace::DescribeTree::NAME.to_owned(),
-            schema: <aether_kinds::trace::DescribeTree as aether_data::Schema>::SCHEMA.clone(),
-        }
-    }
-
-    #[tokio::test]
-    async fn describe_tree_round_trips_observer_reply() {
-        use aether_data::tagged_id::{self, Tag};
-        use aether_data::{MailId, MailboxId, with_tag};
-        use aether_kinds::trace::{DescribeTreeResult, MailNodeWire};
-
-        // Engine record carries the descriptor for the request kind so
-        // the schema-driven encoder in deliver_one knows how to encode
-        // the agent's params into wire bytes. The reply kind doesn't
-        // need a descriptor — the hub decodes via postcard directly.
-        let engines = EngineRegistry::new();
-        let kinds = vec![describe_tree_descriptor()];
-        let (rec, mut rx) = record_with_kinds(0xd001, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        // Sender mailbox we'll use for the root id; only its tag matters
-        // for the JSON wire form. The root has one extra child node.
-        let sender_raw = with_tag(Tag::Mailbox, 0xABCD);
-        let root = MailId {
-            sender: MailboxId(sender_raw),
-            correlation_id: 7,
-        };
-        let child = MailId {
-            sender: MailboxId(with_tag(Tag::Mailbox, 0xCDEF)),
-            correlation_id: 1,
-        };
-
-        // Stub substrate: drain the request the tool sent (we don't
-        // care about its contents for this test — the encode path is
-        // already covered by `encode_schema` tests in aether-codec)
-        // and inject a postcard-encoded `DescribeTreeResult::Ok` reply.
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            let reply = DescribeTreeResult::Ok {
-                root,
-                in_flight: 1,
-                mails: vec![
-                    MailNodeWire {
-                        mail_id: root,
-                        parent: None,
-                        sender: root.sender,
-                        recipient: MailboxId(with_tag(Tag::Mailbox, 1)),
-                        kind: aether_data::KindId(aether_data::with_tag(Tag::Kind, 0xBEEF)),
-                        t_sent: aether_kinds::trace::Nanos(100),
-                        t_received: Some(aether_kinds::trace::Nanos(200)),
-                        t_finished: None,
-                        thread_name: None,
-                    },
-                    MailNodeWire {
-                        mail_id: child,
-                        parent: Some(root),
-                        sender: child.sender,
-                        recipient: MailboxId(with_tag(Tag::Mailbox, 2)),
-                        kind: aether_data::KindId(aether_data::with_tag(Tag::Kind, 0xCAFE)),
-                        t_sent: aether_kinds::trace::Nanos(300),
-                        t_received: None,
-                        t_finished: None,
-                        thread_name: None,
-                    },
-                ],
-            };
-            let payload = postcard::to_allocvec(&reply).expect("encode reply");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind = "aether.trace.describe_tree_result".to_owned();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let remainder = record.replies.try_deliver(&kind, queued);
-            assert!(remainder.is_none(), "reply registry should consume mail");
-        });
-
-        let json = hub
-            .describe_tree(Parameters(args::DescribeTreeArgs {
-                engine_id: id.0.to_string(),
-                root: args::MailIdWire {
-                    sender: tagged_id::encode(sender_raw).unwrap(),
-                    correlation_id: 7,
-                },
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("tool should succeed");
-
-        substrate_task.await.expect("stub substrate task");
-
-        // The reply enum serializes via serde's external tag — `Ok`
-        // becomes a top-level key. Each typed id flows out as a tagged
-        // string per its human-readable serde branch.
-        let raw: serde_json::Value = serde_json::from_str(&json).unwrap();
-        let ok = &raw["Ok"];
-        assert_eq!(ok["in_flight"], 1);
-        let mails = ok["mails"].as_array().unwrap();
-        assert_eq!(mails.len(), 2);
-        // Root node first; its sender renders as a tagged-string mbx-...
-        let m0 = &mails[0];
-        let s0 = m0["sender"].as_str().unwrap();
-        assert!(s0.starts_with("mbx-"), "sender should be tagged: {s0}");
-        let k0 = m0["kind"].as_str().unwrap();
-        assert!(k0.starts_with("knd-"), "kind should be tagged: {k0}");
-        // root field on the Ok variant: composite `{ sender, correlation_id }`
-        let root_obj = &ok["root"];
-        assert!(
-            root_obj["sender"].as_str().unwrap().starts_with("mbx-"),
-            "root.sender should be tagged"
-        );
-        assert_eq!(root_obj["correlation_id"], 7);
-    }
-
-    #[tokio::test]
-    async fn describe_tree_unknown_engine_errors() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(state);
-        let err = hub
-            .describe_tree(Parameters(args::DescribeTreeArgs {
-                engine_id: Uuid::from_u128(0xdead).to_string(),
-                root: args::MailIdWire {
-                    sender: aether_data::tagged_id::encode(aether_data::with_tag(
-                        aether_data::tagged_id::Tag::Mailbox,
-                        1,
-                    ))
-                    .unwrap(),
-                    correlation_id: 0,
-                },
-                timeout_ms: Some(1_000),
-            }))
-            .await
-            .unwrap_err();
-        assert!(format!("{err:?}").contains("unknown engine_id"));
-    }
-
-    /// Issue 735: descriptor for `aether.trace.describe_window`.
-    fn describe_window_descriptor() -> aether_data::KindDescriptor {
-        use aether_data::Kind as _;
-        aether_data::KindDescriptor {
-            name: aether_kinds::trace::DescribeWindow::NAME.to_owned(),
-            schema: <aether_kinds::trace::DescribeWindow as aether_data::Schema>::SCHEMA.clone(),
-        }
-    }
-
-    #[tokio::test]
-    async fn describe_tree_window_round_trips_observer_reply() {
-        use aether_data::tagged_id::Tag;
-        use aether_data::{MailId, MailboxId, with_tag};
-        use aether_kinds::trace::{DescribeWindowResult, MailNodeWire};
-
-        let engines = EngineRegistry::new();
-        let kinds = vec![describe_window_descriptor()];
-        let (rec, mut rx) = record_with_kinds(0xd010, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        let m_in = MailId {
-            sender: MailboxId(with_tag(Tag::Mailbox, 0xAA01)),
-            correlation_id: 1,
-        };
-
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            let reply = DescribeWindowResult::Ok {
-                mails: vec![MailNodeWire {
-                    mail_id: m_in,
-                    parent: None,
-                    sender: m_in.sender,
-                    recipient: MailboxId(with_tag(Tag::Mailbox, 0xAA02)),
-                    kind: aether_data::KindId(aether_data::with_tag(Tag::Kind, 0xCAFE)),
-                    t_sent: aether_kinds::trace::Nanos(500),
-                    t_received: Some(aether_kinds::trace::Nanos(600)),
-                    t_finished: Some(aether_kinds::trace::Nanos(700)),
-                    thread_name: Some("aether-worker-3".to_owned()),
-                }],
-            };
-            let payload = postcard::to_allocvec(&reply).expect("encode reply");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind = "aether.trace.describe_window_result".to_owned();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let remainder = record.replies.try_deliver(&kind, queued);
-            assert!(remainder.is_none(), "reply registry should consume mail");
-        });
-
-        let json = hub
-            .describe_tree_window(Parameters(args::DescribeTreeWindowArgs {
-                engine_id: id.0.to_string(),
-                window: args::TraceWindowArg::Relative { last_ms: 1_000 },
-                max_mails: Some(100),
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("tool should succeed");
-
-        substrate_task.await.expect("stub substrate task");
-
-        let raw: serde_json::Value = serde_json::from_str(&json).unwrap();
-        let ok = &raw["Ok"];
-        let mails = ok["mails"].as_array().unwrap();
-        assert_eq!(mails.len(), 1);
-        let m = &mails[0];
-        assert!(
-            m["sender"].as_str().unwrap().starts_with("mbx-"),
-            "sender should be tagged"
-        );
-        assert!(
-            m["kind"].as_str().unwrap().starts_with("knd-"),
-            "kind should be tagged"
-        );
-        assert_eq!(m["t_sent"], 500);
-        assert_eq!(m["thread_name"], "aether-worker-3");
-    }
-
-    #[tokio::test]
-    async fn describe_tree_window_too_many_surfaces_err() {
-        use aether_kinds::trace::DescribeWindowResult;
-
-        let engines = EngineRegistry::new();
-        let kinds = vec![describe_window_descriptor()];
-        let (rec, mut rx) = record_with_kinds(0xd011, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        // Stub substrate replies with Err::too_many — the hub should
-        // pass it through as the externally-tagged JSON envelope so
-        // the agent can read the matched count.
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            let reply = DescribeWindowResult::Err {
-                too_many: Some(12_345),
-            };
-            let payload = postcard::to_allocvec(&reply).expect("encode reply");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind = "aether.trace.describe_window_result".to_owned();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let remainder = record.replies.try_deliver(&kind, queued);
-            assert!(remainder.is_none(), "reply registry should consume mail");
-        });
-
-        let json = hub
-            .describe_tree_window(Parameters(args::DescribeTreeWindowArgs {
-                engine_id: id.0.to_string(),
-                window: args::TraceWindowArg::Absolute {
-                    start_ns: 0,
-                    end_ns: Some(1_000_000_000_000),
-                },
-                max_mails: Some(5),
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("tool should succeed");
-
-        substrate_task.await.expect("stub substrate task");
-
-        let raw: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert_eq!(raw["Err"]["too_many"], 12_345);
-    }
-
-    #[tokio::test]
-    async fn describe_tree_window_unknown_engine_errors() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(state);
-        let err = hub
-            .describe_tree_window(Parameters(args::DescribeTreeWindowArgs {
-                engine_id: Uuid::from_u128(0xdead_beef).to_string(),
-                window: args::TraceWindowArg::Relative { last_ms: 100 },
-                max_mails: None,
-                timeout_ms: Some(1_000),
-            }))
-            .await
-            .unwrap_err();
-        assert!(format!("{err:?}").contains("unknown engine_id"));
-    }
-
-    fn list_active_roots_descriptor() -> aether_data::KindDescriptor {
-        use aether_data::Kind as _;
-        aether_data::KindDescriptor {
-            name: aether_kinds::trace::ListActiveRoots::NAME.to_owned(),
-            schema: <aether_kinds::trace::ListActiveRoots as aether_data::Schema>::SCHEMA.clone(),
-        }
-    }
-
-    #[tokio::test]
-    async fn list_active_roots_round_trips_observer_reply() {
-        use aether_data::tagged_id::Tag;
-        use aether_data::{MailId, MailboxId, with_tag};
-        use aether_kinds::trace::{ListActiveRootsResult, RootSummaryWire};
-
-        let engines = EngineRegistry::new();
-        let kinds = vec![list_active_roots_descriptor()];
-        let (rec, mut rx) = record_with_kinds(0xd003, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        let root_a = MailId {
-            sender: MailboxId(with_tag(Tag::Mailbox, 0x111)),
-            correlation_id: 1,
-        };
-        let root_b = MailId {
-            sender: MailboxId(with_tag(Tag::Mailbox, 0x222)),
-            correlation_id: 2,
-        };
-
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            let reply = ListActiveRootsResult {
-                roots: vec![
-                    RootSummaryWire {
-                        root: root_b,
-                        kind: aether_data::KindId(with_tag(Tag::Kind, 0xCAFE)),
-                        sender: root_b.sender,
-                        recipient: MailboxId(with_tag(Tag::Mailbox, 9)),
-                        t_sent: aether_kinds::trace::Nanos(2_000_000_000),
-                        in_flight: 0,
-                    },
-                    RootSummaryWire {
-                        root: root_a,
-                        kind: aether_data::KindId(with_tag(Tag::Kind, 0xBEEF)),
-                        sender: root_a.sender,
-                        recipient: MailboxId(with_tag(Tag::Mailbox, 8)),
-                        t_sent: aether_kinds::trace::Nanos(1_000_000_000),
-                        in_flight: 3,
-                    },
-                ],
-            };
-            let payload = postcard::to_allocvec(&reply).expect("encode");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind = "aether.trace.list_active_roots_result".to_owned();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let remainder = record.replies.try_deliver(&kind, queued);
-            assert!(remainder.is_none(), "reply registry should consume mail");
-        });
-
-        let json = hub
-            .list_active_roots(Parameters(args::ListActiveRootsArgs {
-                engine_id: id.0.to_string(),
-                since_ms: Some(60_000),
-                max: Some(50),
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("tool should succeed");
-
-        substrate_task.await.expect("stub substrate task");
-
-        let raw: serde_json::Value = serde_json::from_str(&json).unwrap();
-        let roots = raw["roots"].as_array().unwrap();
-        assert_eq!(roots.len(), 2);
-        // Order preserved from the substrate's reply (the substrate
-        // sorts desc by t_sent before encoding).
-        assert_eq!(roots[0]["in_flight"], 0);
-        assert_eq!(roots[1]["in_flight"], 3);
-        // Each typed id renders as its tagged-string form.
-        let r0_sender = roots[0]["sender"].as_str().unwrap();
-        assert!(r0_sender.starts_with("mbx-"), "sender tagged: {r0_sender}");
-        let r0_kind = roots[0]["kind"].as_str().unwrap();
-        assert!(r0_kind.starts_with("knd-"), "kind tagged: {r0_kind}");
-        // Composite root field.
-        assert!(
-            roots[0]["root"]["sender"]
-                .as_str()
-                .unwrap()
-                .starts_with("mbx-"),
-            "root.sender tagged"
-        );
-    }
-
-    #[tokio::test]
-    async fn list_active_roots_unknown_engine_errors() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(state);
-        let err = hub
-            .list_active_roots(Parameters(args::ListActiveRootsArgs {
-                engine_id: Uuid::from_u128(0xdead).to_string(),
-                since_ms: None,
-                max: None,
-                timeout_ms: Some(1_000),
-            }))
-            .await
-            .unwrap_err();
-        assert!(format!("{err:?}").contains("unknown engine_id"));
-    }
-
-    #[tokio::test]
-    async fn describe_tree_rejects_malformed_sender() {
-        let engines = EngineRegistry::new();
-        let (rec, _rx) = record(0xd0c2);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-        let err = hub
-            .describe_tree(Parameters(args::DescribeTreeArgs {
-                engine_id: id.0.to_string(),
-                root: args::MailIdWire {
-                    sender: "not-a-mailbox".into(),
-                    correlation_id: 0,
-                },
-                timeout_ms: Some(1_000),
-            }))
-            .await
-            .unwrap_err();
-        assert!(
-            format!("{err:?}").to_lowercase().contains("root.sender"),
-            "{err:?}"
-        );
-    }
-
-    /// Issue 728: dump_trace calls the same describe_tree path the
-    /// existing tool uses, then re-shapes the reply as trace JSON.
-    /// The output_path branch writes to disk and returns
-    /// `{"path": ...}`; the inline branch returns the JSON.
-    #[tokio::test]
-    async fn dump_trace_writes_file_when_path_set() {
-        use aether_data::tagged_id::{self, Tag};
-        use aether_data::{MailId, MailboxId, with_tag};
-        use aether_kinds::trace::{DescribeTreeResult, MailNodeWire};
-
-        let engines = EngineRegistry::new();
-        let kinds = vec![describe_tree_descriptor()];
-        let (rec, mut rx) = record_with_kinds(0xd728, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        let sender_raw = with_tag(Tag::Mailbox, 0xCAFE);
-        let root = MailId {
-            sender: MailboxId(sender_raw),
-            correlation_id: 1,
-        };
-
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            let reply = DescribeTreeResult::Ok {
-                root,
-                in_flight: 0,
-                mails: vec![MailNodeWire {
-                    mail_id: root,
-                    parent: None,
-                    sender: root.sender,
-                    recipient: MailboxId(with_tag(Tag::Mailbox, 0xBEEF)),
-                    kind: aether_data::KindId(with_tag(Tag::Kind, 0xDEAD)),
-                    t_sent: aether_kinds::trace::Nanos(1_000),
-                    t_received: Some(aether_kinds::trace::Nanos(2_000)),
-                    t_finished: Some(aether_kinds::trace::Nanos(5_000)),
-                    thread_name: None,
-                }],
-            };
-            let payload = postcard::to_allocvec(&reply).expect("encode reply");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind = "aether.trace.describe_tree_result".to_owned();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let remainder = record.replies.try_deliver(&kind, queued);
-            assert!(remainder.is_none(), "reply registry should consume mail");
-        });
-
-        let tmp =
-            std::env::temp_dir().join(format!("aether-trace-728-{}.json", std::process::id()));
-        let response = hub
-            .dump_trace(Parameters(args::DumpTraceArgs {
-                engine_id: id.0.to_string(),
-                root: args::MailIdWire {
-                    sender: tagged_id::encode(sender_raw).unwrap(),
-                    correlation_id: 1,
-                },
-                output_path: Some(tmp.to_string_lossy().into_owned()),
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("tool should succeed");
-        substrate_task.await.expect("stub substrate task");
-
-        let parsed: serde_json::Value = serde_json::from_str(&response).unwrap();
-        assert_eq!(parsed["path"].as_str().unwrap(), tmp.to_string_lossy());
-
-        let written = std::fs::read_to_string(&tmp).expect("trace file written");
-        let trace: serde_json::Value = serde_json::from_str(&written).unwrap();
-        let events = trace["traceEvents"].as_array().unwrap();
-        // Filter out the auto-prepended `process_name` metadata
-        // events (issue 731 follow-up to silence Perfetto's hashed-pid
-        // display) so the assertion stays focused on the slice shape.
-        let slices: Vec<&serde_json::Value> = events.iter().filter(|e| e["ph"] != "M").collect();
-        assert_eq!(slices.len(), 1, "one mail -> one X event");
-        assert_eq!(slices[0]["ph"], "X");
-
-        std::fs::remove_file(&tmp).ok();
-    }
-
-    /// Inline branch: when `output_path` is omitted, the trace JSON
-    /// flows back as the tool result.
-    #[tokio::test]
-    async fn dump_trace_returns_inline_when_path_unset() {
-        use aether_data::tagged_id::{self, Tag};
-        use aether_data::{MailId, MailboxId, with_tag};
-        use aether_kinds::trace::{DescribeTreeResult, MailNodeWire};
-
-        let engines = EngineRegistry::new();
-        let kinds = vec![describe_tree_descriptor()];
-        let (rec, mut rx) = record_with_kinds(0xd729, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        let sender_raw = with_tag(Tag::Mailbox, 0xCAFE);
-        let root = MailId {
-            sender: MailboxId(sender_raw),
-            correlation_id: 1,
-        };
-
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            let reply = DescribeTreeResult::Ok {
-                root,
-                in_flight: 0,
-                mails: vec![MailNodeWire {
-                    mail_id: root,
-                    parent: None,
-                    sender: root.sender,
-                    recipient: MailboxId(with_tag(Tag::Mailbox, 0xBEEF)),
-                    kind: aether_data::KindId(with_tag(Tag::Kind, 0xDEAD)),
-                    t_sent: aether_kinds::trace::Nanos(1_000),
-                    t_received: Some(aether_kinds::trace::Nanos(2_000)),
-                    t_finished: Some(aether_kinds::trace::Nanos(5_000)),
-                    thread_name: None,
-                }],
-            };
-            let payload = postcard::to_allocvec(&reply).expect("encode reply");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind = "aether.trace.describe_tree_result".to_owned();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let _ = record.replies.try_deliver(&kind, queued);
-        });
-
-        let response = hub
-            .dump_trace(Parameters(args::DumpTraceArgs {
-                engine_id: id.0.to_string(),
-                root: args::MailIdWire {
-                    sender: tagged_id::encode(sender_raw).unwrap(),
-                    correlation_id: 1,
-                },
-                output_path: None,
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("tool should succeed");
-        substrate_task.await.expect("stub substrate task");
-
-        let trace: serde_json::Value = serde_json::from_str(&response).unwrap();
-        let events = trace["traceEvents"].as_array().unwrap();
-        let slices: Vec<&serde_json::Value> = events.iter().filter(|e| e["ph"] != "M").collect();
-        assert_eq!(slices.len(), 1);
-        assert_eq!(slices[0]["ph"], "X");
-    }
-
-    /// Issue 735 phase 3: dump_trace_window calls describe_window then
-    /// renders the resulting flat mail set as Chrome trace JSON. With
-    /// `output_path` set, the trace is written to disk and the tool
-    /// returns `{"path": ...}` — same shape as `dump_trace`. The
-    /// rendered events come straight from the windowed mails (no root
-    /// context); each mail with `t_received`/`t_finished` populated
-    /// becomes one `ph: "X"` slice.
-    #[tokio::test]
-    async fn dump_trace_window_writes_file_when_path_set() {
-        use aether_data::tagged_id::Tag;
-        use aether_data::{MailId, MailboxId, with_tag};
-        use aether_kinds::trace::{DescribeWindowResult, MailNodeWire};
-
-        let engines = EngineRegistry::new();
-        let kinds = vec![describe_window_descriptor()];
-        let (rec, mut rx) = record_with_kinds(0xd735, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        // Two unrelated mails (different roots) inside the window —
-        // dump_trace_window renders both even though they're in
-        // separate chains.
-        let m_a = MailId {
-            sender: MailboxId(with_tag(Tag::Mailbox, 0xAA)),
-            correlation_id: 1,
-        };
-        let m_b = MailId {
-            sender: MailboxId(with_tag(Tag::Mailbox, 0xBB)),
-            correlation_id: 1,
-        };
-
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            let reply = DescribeWindowResult::Ok {
-                mails: vec![
-                    MailNodeWire {
-                        mail_id: m_a,
-                        parent: None,
-                        sender: m_a.sender,
-                        recipient: MailboxId(with_tag(Tag::Mailbox, 1)),
-                        kind: aether_data::KindId(with_tag(Tag::Kind, 0xCAFE)),
-                        t_sent: aether_kinds::trace::Nanos(1_000),
-                        t_received: Some(aether_kinds::trace::Nanos(2_000)),
-                        t_finished: Some(aether_kinds::trace::Nanos(3_000)),
-                        thread_name: None,
-                    },
-                    MailNodeWire {
-                        mail_id: m_b,
-                        parent: None,
-                        sender: m_b.sender,
-                        recipient: MailboxId(with_tag(Tag::Mailbox, 2)),
-                        kind: aether_data::KindId(with_tag(Tag::Kind, 0xDEAD)),
-                        t_sent: aether_kinds::trace::Nanos(5_000),
-                        t_received: Some(aether_kinds::trace::Nanos(6_000)),
-                        t_finished: Some(aether_kinds::trace::Nanos(8_000)),
-                        thread_name: None,
-                    },
-                ],
-            };
-            let payload = postcard::to_allocvec(&reply).expect("encode reply");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind = "aether.trace.describe_window_result".to_owned();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let remainder = record.replies.try_deliver(&kind, queued);
-            assert!(remainder.is_none(), "reply registry should consume mail");
-        });
-
-        let tmp =
-            std::env::temp_dir().join(format!("aether-trace-window-{}.json", std::process::id()));
-        let response = hub
-            .dump_trace_window(Parameters(args::DumpTraceWindowArgs {
-                engine_id: id.0.to_string(),
-                window: args::TraceWindowArg::Relative { last_ms: 10_000 },
-                output_path: Some(tmp.to_string_lossy().into_owned()),
-                max_mails: Some(1_000),
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("tool should succeed");
-        substrate_task.await.expect("stub substrate task");
-
-        let parsed: serde_json::Value = serde_json::from_str(&response).unwrap();
-        assert_eq!(parsed["path"].as_str().unwrap(), tmp.to_string_lossy());
-
-        let written = std::fs::read_to_string(&tmp).expect("trace file written");
-        let trace: serde_json::Value = serde_json::from_str(&written).unwrap();
-        let events = trace["traceEvents"].as_array().unwrap();
-        let slices: Vec<&serde_json::Value> = events.iter().filter(|e| e["ph"] != "M").collect();
-        assert_eq!(slices.len(), 2, "two mails -> two X events");
-        for s in &slices {
-            assert_eq!(s["ph"], "X");
-        }
-
-        std::fs::remove_file(&tmp).ok();
-    }
-
-    /// Issue 735 phase 3: when the substrate replies `Err::too_many`
-    /// the hub surfaces it as an `invalid_params` MCP error mentioning
-    /// the matched count, so the agent knows whether to narrow the
-    /// window or raise `max_mails`. No file is written.
-    #[tokio::test]
-    async fn dump_trace_window_too_many_surfaces_error() {
-        use aether_kinds::trace::DescribeWindowResult;
-
-        let engines = EngineRegistry::new();
-        let kinds = vec![describe_window_descriptor()];
-        let (rec, mut rx) = record_with_kinds(0xd736, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            let reply = DescribeWindowResult::Err {
-                too_many: Some(54_321),
-            };
-            let payload = postcard::to_allocvec(&reply).expect("encode reply");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind = "aether.trace.describe_window_result".to_owned();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let remainder = record.replies.try_deliver(&kind, queued);
-            assert!(remainder.is_none(), "reply registry should consume mail");
-        });
-
-        let err = hub
-            .dump_trace_window(Parameters(args::DumpTraceWindowArgs {
-                engine_id: id.0.to_string(),
-                window: args::TraceWindowArg::Relative { last_ms: 10_000 },
-                output_path: None,
-                max_mails: Some(5),
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .unwrap_err();
-        substrate_task.await.expect("stub substrate task");
-
-        let msg = format!("{err:?}");
-        assert!(
-            msg.contains("54321") && msg.contains("max_mails"),
-            "error should mention matched count + how to fix: {msg}"
-        );
-    }
-
-    #[tokio::test]
-    async fn describe_component_rejects_malformed_mailbox_id() {
-        // ADR-0064: a bare-number mailbox id is no longer valid wire
-        // form. Reject at the boundary with a typed error so the
-        // agent sees a clear "looks malformed" signal rather than a
-        // mysterious lookup miss.
-        let engines = EngineRegistry::new();
-        let (rec, _rx) = record(52);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-
-        let err = hub
-            .describe_component(Parameters(args::DescribeComponentArgs {
-                engine_id: id.0.to_string(),
-                mailbox_id: "999".into(),
-            }))
-            .await
-            .unwrap_err();
-        assert!(format!("{err:?}").to_lowercase().contains("mailbox_id"));
-    }
-
-    #[tokio::test]
-    async fn describe_component_rejects_wrong_tag() {
-        // Passing a kind id (`knd-...`) where a mailbox id is expected
-        // is exactly what the tag bits exist to catch. Verify the
-        // boundary surfaces it as a tag-mismatch error rather than
-        // silently treating the kind id as a mailbox id and missing
-        // the lookup.
-        use aether_data::tagged_id::{self, Tag};
-        use aether_data::with_tag;
-
-        let engines = EngineRegistry::new();
-        let (rec, _rx) = record(53);
-        let id = rec.id;
-        engines.insert(rec);
-        let hub = Hub::new(test_state(engines, SessionRegistry::new()));
-
-        let kind_form = tagged_id::encode(with_tag(Tag::Kind, 42)).unwrap();
-        let err = hub
-            .describe_component(Parameters(args::DescribeComponentArgs {
-                engine_id: id.0.to_string(),
-                mailbox_id: kind_form,
-            }))
-            .await
-            .unwrap_err();
-        assert!(format!("{err:?}").contains("tag mismatch"));
-    }
-
-    async fn push_queued(sessions: &SessionRegistry, token: SessionToken, mail: QueuedMail) {
-        sessions
-            .get(&token)
-            .expect("session")
-            .mail_tx
-            .send(mail)
-            .await
-            .expect("send");
-    }
-
-    fn queued(engine: u128, kind: &str, payload: Vec<u8>, broadcast: bool) -> QueuedMail {
-        QueuedMail {
-            engine_id: EngineId(Uuid::from_u128(engine)),
-            kind_name: kind.into(),
-            payload,
-            broadcast,
-            origin: None,
-        }
-    }
-
-    async fn drain(hub: &Hub, max: Option<u32>) -> Vec<ReceivedMail> {
-        let json = hub
-            .receive_mail(Parameters(ReceiveMailArgs { max }))
-            .await
-            .unwrap();
-        serde_json::from_str(&json).unwrap()
-    }
-
-    #[tokio::test]
-    async fn receive_mail_empty_queue_returns_empty_array() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(state);
-        let got = drain(&hub, None).await;
-        assert!(got.is_empty());
-    }
-
-    #[tokio::test]
-    async fn receive_mail_drains_everything_by_default() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let token = hub.session.token;
-
-        push_queued(
-            &state.sessions,
-            token,
-            queued(7, "aether.observation.ping", vec![1, 2], false),
-        )
-        .await;
-        push_queued(
-            &state.sessions,
-            token,
-            queued(7, "aether.observation.world", vec![9], true),
-        )
-        .await;
-
-        let got = drain(&hub, None).await;
-        assert_eq!(got.len(), 2);
-        assert_eq!(got[0].kind_name, "aether.observation.ping");
-        assert_eq!(got[0].payload_bytes, vec![1, 2]);
-        assert!(!got[0].broadcast);
-        assert_eq!(got[0].engine_id, Uuid::from_u128(7).to_string());
-        assert!(got[0].origin.is_none());
-        assert!(got[1].broadcast);
-        assert!(got[1].origin.is_none());
-
-        // Queue is now empty.
-        let got = drain(&hub, None).await;
-        assert!(got.is_empty());
-    }
-
-    #[tokio::test]
-    async fn receive_mail_surfaces_origin() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let token = hub.session.token;
-
-        let mut with_origin = queued(3, "aether.observation.frame_stats", vec![], true);
-        with_origin.origin = Some("render".into());
-        push_queued(&state.sessions, token, with_origin).await;
-
-        let got = drain(&hub, None).await;
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].origin.as_deref(), Some("render"));
-    }
-
-    #[tokio::test]
-    async fn receive_mail_respects_max() {
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let token = hub.session.token;
-        for i in 0..5u8 {
-            push_queued(
-                &state.sessions,
-                token,
-                queued(1, "aether.tick", vec![i], false),
-            )
-            .await;
-        }
-
-        let first = drain(&hub, Some(2)).await;
-        assert_eq!(first.len(), 2);
-        let second = drain(&hub, Some(2)).await;
-        assert_eq!(second.len(), 2);
-        let rest = drain(&hub, None).await;
-        assert_eq!(rest.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn receive_mail_decodes_params_against_descriptor() {
-        // FrameStats-shaped: cast struct with two u64 fields. The
-        // engine ships raw cast bytes; the hub looks up the descriptor
-        // and lifts them into structured `params`.
-        use aether_data::{KindDescriptor, NamedField, Primitive, SchemaType};
-        let engines = EngineRegistry::new();
-        let kinds = vec![KindDescriptor {
-            name: "aether.observation.frame_stats".into(),
-            schema: SchemaType::Struct {
-                repr_c: true,
-                fields: vec![
-                    NamedField {
-                        name: "frame".into(),
-                        ty: SchemaType::Scalar(Primitive::U64),
-                    },
-                    NamedField {
-                        name: "triangles".into(),
-                        ty: SchemaType::Scalar(Primitive::U64),
-                    },
-                ]
-                .into(),
-            },
-        }];
-        let (rec, _rx) = record_with_kinds(33, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let token = hub.session.token;
-
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&120u64.to_le_bytes());
-        payload.extend_from_slice(&7u64.to_le_bytes());
-        push_queued(
-            &state.sessions,
-            token,
-            QueuedMail {
-                engine_id: id,
-                kind_name: "aether.observation.frame_stats".into(),
-                payload,
-                broadcast: true,
-                origin: None,
-            },
-        )
-        .await;
-
-        let got = drain(&hub, None).await;
-        assert_eq!(got.len(), 1);
-        let item = &got[0];
-        assert_eq!(
-            item.params,
-            Some(serde_json::json!({"frame": 120u64, "triangles": 7u64}))
-        );
-        assert!(item.decode_error.is_none());
-        // payload_bytes still populated alongside params (escape hatch).
-        assert_eq!(item.payload_bytes.len(), 16);
-    }
-
-    #[tokio::test]
-    async fn receive_mail_decode_failure_populates_error() {
-        // Descriptor declares 8-byte u64; hub gets only 2 bytes.
-        // Decoder must surface a Truncated error in `decode_error` and
-        // leave `params` null without dropping the item.
-        use aether_data::{KindDescriptor, NamedField, Primitive, SchemaType};
-        let engines = EngineRegistry::new();
-        let kinds = vec![KindDescriptor {
-            name: "demo.short".into(),
-            schema: SchemaType::Struct {
-                repr_c: true,
-                fields: vec![NamedField {
-                    name: "n".into(),
-                    ty: SchemaType::Scalar(Primitive::U64),
-                }]
-                .into(),
-            },
-        }];
-        let (rec, _rx) = record_with_kinds(34, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let token = hub.session.token;
-
-        push_queued(
-            &state.sessions,
-            token,
-            QueuedMail {
-                engine_id: id,
-                kind_name: "demo.short".into(),
-                payload: vec![1, 2],
-                broadcast: false,
-                origin: None,
-            },
-        )
-        .await;
-
-        let got = drain(&hub, None).await;
-        assert_eq!(got.len(), 1);
-        let item = &got[0];
-        assert!(item.params.is_none());
-        let err = item
-            .decode_error
-            .as_deref()
-            .expect("decode error populated");
-        assert!(err.contains("truncated"), "got: {err}");
-        assert_eq!(item.payload_bytes, vec![1, 2]);
-    }
-
-    #[tokio::test]
-    async fn receive_mail_unknown_kind_falls_back_to_bytes() {
-        // Engine sent a kind it never declared at handshake (or the
-        // descriptor was lost). Decode reports the missing descriptor;
-        // bytes survive for the agent to inspect.
-        let engines = EngineRegistry::new();
-        let (rec, _rx) = record_with_kinds(35, vec![]);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let token = hub.session.token;
-
-        push_queued(
-            &state.sessions,
-            token,
-            QueuedMail {
-                engine_id: id,
-                kind_name: "demo.unknown".into(),
-                payload: vec![9, 9, 9],
-                broadcast: false,
-                origin: None,
-            },
-        )
-        .await;
-
-        let got = drain(&hub, None).await;
-        assert_eq!(got.len(), 1);
-        assert!(got[0].params.is_none());
-        let err = got[0].decode_error.as_deref().unwrap();
-        assert!(err.contains("no descriptor"), "got: {err}");
-        assert_eq!(got[0].payload_bytes, vec![9, 9, 9]);
-    }
-
-    #[tokio::test]
-    async fn receive_mail_unit_kind_decodes_to_null_params() {
-        use aether_data::{KindDescriptor, SchemaType};
-        let engines = EngineRegistry::new();
-        let kinds = vec![KindDescriptor {
-            name: "aether.observation.ping".into(),
-            schema: SchemaType::Unit,
-        }];
-        let (rec, _rx) = record_with_kinds(36, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let token = hub.session.token;
-
-        push_queued(
-            &state.sessions,
-            token,
-            QueuedMail {
-                engine_id: id,
-                kind_name: "aether.observation.ping".into(),
-                payload: vec![],
-                broadcast: false,
-                origin: None,
-            },
-        )
-        .await;
-
-        // Read raw JSON rather than going through `drain` — `Option<Value>`
-        // collapses `Some(Null)` and `None` over deserialization, so the
-        // distinction between "decoded to null" and "no value" is only
-        // observable in the wire form. The MCP client sees the wire form.
-        let json = hub
-            .receive_mail(Parameters(ReceiveMailArgs { max: None }))
-            .await
-            .unwrap();
-        let raw: serde_json::Value = serde_json::from_str(&json).unwrap();
-        let item = &raw.as_array().unwrap()[0];
-        assert_eq!(item.get("params"), Some(&serde_json::Value::Null));
-        // decode_error skipped on success thanks to `skip_serializing_if`.
-        assert!(item.get("decode_error").is_none());
-    }
-
-    #[tokio::test]
-    async fn receive_mail_scoped_to_own_session() {
-        // Push into session A's queue; session B's drain should see nothing.
-        let state = test_state(EngineRegistry::new(), SessionRegistry::new());
-        let hub_a = Hub::new(Arc::clone(&state));
-        let hub_b = Hub::new(Arc::clone(&state));
-
-        push_queued(
-            &state.sessions,
-            hub_a.session.token,
-            queued(1, "aether.tick", vec![42], false),
-        )
-        .await;
-
-        let got_b = drain(&hub_b, None).await;
-        assert!(got_b.is_empty());
-
-        let got_a = drain(&hub_a, None).await;
-        assert_eq!(got_a.len(), 1);
-        assert_eq!(got_a[0].payload_bytes, vec![42]);
-    }
-
-    /// Descriptor for `aether.render.capture_frame` that matches
-    /// the real substrate's schema: a postcard struct with `mails`
-    /// and `after_mails` fields, both `Vec<MailEnvelope>`. Used
-    /// across capture_frame tests so they all exercise the same
-    /// wire shape.
-    fn capture_frame_kind_descriptor() -> aether_data::KindDescriptor {
-        use aether_data::{KindDescriptor, NamedField, Primitive, SchemaCell, SchemaType};
-        let envelope = SchemaType::Struct {
-            repr_c: false,
-            fields: vec![
-                NamedField {
-                    name: "recipient_name".into(),
-                    ty: SchemaType::String,
-                },
-                NamedField {
-                    name: "kind_name".into(),
-                    ty: SchemaType::String,
-                },
-                NamedField {
-                    name: "payload".into(),
-                    ty: SchemaType::Bytes,
-                },
-                NamedField {
-                    name: "count".into(),
-                    ty: SchemaType::Scalar(Primitive::U32),
-                },
-            ]
-            .into(),
-        };
-        KindDescriptor {
-            name: "aether.render.capture_frame".into(),
-            schema: SchemaType::Struct {
-                repr_c: false,
-                fields: vec![
-                    NamedField {
-                        name: "mails".into(),
-                        ty: SchemaType::Vec(SchemaCell::owned(envelope.clone())),
-                    },
-                    NamedField {
-                        name: "after_mails".into(),
-                        ty: SchemaType::Vec(SchemaCell::owned(envelope)),
-                    },
-                ]
-                .into(),
-            },
-        }
-    }
-
-    #[tokio::test]
-    async fn capture_frame_returns_image_on_substrate_reply() {
-        // End-to-end of the MCP-side plumbing with a stubbed "substrate":
-        // the tool sends request mail (which we drain off the engine's
-        // mail_tx and ignore), then we inject a postcard-encoded
-        // `Ok { png }` reply through the session's reply registry. The
-        // tool should return a `CallToolResult` carrying the PNG as
-        // image content.
-
-        let engines = EngineRegistry::new();
-        let kinds = vec![capture_frame_kind_descriptor()];
-        let (rec, mut rx) = record_with_kinds(777, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-
-        // Stub substrate: wait until the tool has sent its request
-        // (which means the reply registration is already installed),
-        // then push the reply through the session's mail tx path.
-        let sessions_clone = state.sessions.clone();
-        let substrate_task = tokio::spawn(async move {
-            // Drain the request the tool sent to the engine.
-            let _req = rx.recv().await.expect("tool should send request");
-            // Build the reply payload — postcard-encoded
-            // `CaptureFrameResult::Ok { png: vec![PNG...] }`. Use the
-            // local wire mirror so this test doesn't depend on
-            // aether-kinds being present.
-            #[derive(serde::Serialize)]
-            enum ReplyMirror {
-                #[allow(dead_code)]
-                Ok { png: Vec<u8> },
-                #[allow(dead_code)]
-                Err { error: String },
-            }
-            let fake_png = vec![0x89, b'P', b'N', b'G', 0x0d, 0x0a, 0x1a, 0x0a, 1, 2, 3, 4];
-            let payload = postcard::to_allocvec(&ReplyMirror::Ok {
-                png: fake_png.clone(),
-            })
-            .expect("encode");
-            // Push as if it were a `ClaudeAddress::Session(token)`
-            // mail landing on this session. The await-reply diversion
-            // runs inside the engine_reader's `try_deliver`; in tests
-            // we short-circuit by driving the diversion directly.
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind: String = "aether.render.capture_frame_result".into();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            let remainder = record.replies.try_deliver(&kind, queued);
-            assert!(
-                remainder.is_none(),
-                "reply registry should have consumed the mail; tool wasn't awaiting"
-            );
-        });
-
-        let result = hub
-            .capture_frame(Parameters(CaptureFrameArgs {
-                engine_id: id.0.to_string(),
-                mails: vec![],
-                after_mails: vec![],
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("tool should succeed");
-
-        substrate_task.await.expect("stub substrate task");
-
-        assert_eq!(result.content.len(), 1);
-        let image = result.content[0]
-            .as_image()
-            .expect("content should be image");
-        assert_eq!(image.mime_type, "image/png");
-        let decoded = base64::engine::general_purpose::STANDARD
-            .decode(&image.data)
-            .expect("valid base64");
-        assert_eq!(
-            decoded,
-            vec![0x89, b'P', b'N', b'G', 0x0d, 0x0a, 0x1a, 0x0a, 1, 2, 3, 4]
-        );
-    }
-
-    #[tokio::test]
-    async fn capture_frame_surfaces_substrate_err_variant() {
-        let engines = EngineRegistry::new();
-        let kinds = vec![capture_frame_kind_descriptor()];
-        let (rec, mut rx) = record_with_kinds(778, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-        let sessions_clone = state.sessions.clone();
-
-        let substrate_task = tokio::spawn(async move {
-            let _req = rx.recv().await.expect("tool should send request");
-            #[derive(serde::Serialize)]
-            enum ReplyMirror {
-                #[allow(dead_code)]
-                Ok {
-                    png: Vec<u8>,
-                },
-                Err {
-                    error: String,
-                },
-            }
-            let payload = postcard::to_allocvec(&ReplyMirror::Err {
-                error: "gpu lost".into(),
-            })
-            .expect("encode");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind: String = "aether.render.capture_frame_result".into();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload,
-                broadcast: false,
-                origin: None,
-            };
-            record.replies.try_deliver(&kind, queued);
-        });
-
-        let err = hub
-            .capture_frame(Parameters(CaptureFrameArgs {
-                engine_id: id.0.to_string(),
-                mails: vec![],
-                after_mails: vec![],
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .unwrap_err();
-
-        substrate_task.await.expect("stub substrate task");
-
-        let msg = format!("{err:?}");
-        assert!(msg.contains("gpu lost"), "unexpected error: {msg}");
-    }
-
-    #[tokio::test]
-    async fn capture_frame_rejects_second_concurrent_call() {
-        let engines = EngineRegistry::new();
-        let kinds = vec![capture_frame_kind_descriptor()];
-        let (rec, _rx) = record_with_kinds(779, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-
-        // Pre-register a waiter for capture_frame_result via the same
-        // path the tool uses, so the second call sees a conflict
-        // without us actually racing two tool calls.
-        let (_guard, _rx) = hub
-            .session
-            .replies
-            .register("aether.render.capture_frame_result".into())
-            .expect("first registration");
-
-        let err = hub
-            .capture_frame(Parameters(CaptureFrameArgs {
-                engine_id: id.0.to_string(),
-                mails: vec![],
-                after_mails: vec![],
-                timeout_ms: Some(100),
-            }))
-            .await
-            .unwrap_err();
-        let msg = format!("{err:?}");
-        assert!(
-            msg.contains("already in flight"),
-            "expected conflict: {msg}"
-        );
-    }
-
-    #[tokio::test]
-    async fn capture_frame_times_out_when_no_reply() {
-        let engines = EngineRegistry::new();
-        let kinds = vec![capture_frame_kind_descriptor()];
-        let (rec, mut rx) = record_with_kinds(780, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-
-        // Drain the request the tool will send so its mail_tx doesn't
-        // back-pressure, but never reply. The timeout fires.
-        let drain = tokio::spawn(async move {
-            let _ = rx.recv().await;
-        });
-
-        let err = hub
-            .capture_frame(Parameters(CaptureFrameArgs {
-                engine_id: id.0.to_string(),
-                mails: vec![],
-                after_mails: vec![],
-                timeout_ms: Some(50),
-            }))
-            .await
-            .unwrap_err();
-        drain.await.ok();
-        let msg = format!("{err:?}");
-        assert!(msg.contains("timed out"), "expected timeout: {msg}");
-    }
-
-    #[tokio::test]
-    async fn capture_frame_encodes_bundle_into_request_payload() {
-        // Verify the bundle path: tool takes a `mails` array, resolves
-        // each MailSpec via its kind descriptor into bytes, wraps
-        // into a CaptureFrame, and the request-mail payload on the
-        // wire carries a valid postcard-encoded bundle the substrate
-        // could decode.
-        use aether_data::{KindDescriptor, SchemaType};
-        let engines = EngineRegistry::new();
-        // Two kinds on this engine: capture_frame plus a `demo.tick`
-        // Unit kind we'll bundle into the capture request.
-        let kinds = vec![
-            capture_frame_kind_descriptor(),
-            KindDescriptor {
-                name: "demo.tick".into(),
-                schema: SchemaType::Unit,
-            },
-        ];
-        let (rec, mut rx) = record_with_kinds(781, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(Arc::clone(&state));
-        let session_token = hub.session.token;
-        let sessions_clone = state.sessions.clone();
-
-        let substrate_task = tokio::spawn(async move {
-            let req = rx.recv().await.expect("tool should send request");
-            // Decode the CaptureFrame the tool built — a postcard
-            // struct with a `Vec<MailEnvelope>` field. Using a local
-            // mirror keeps this test in the hub crate.
-            #[derive(serde::Deserialize, Debug)]
-            struct EnvelopeMirror {
-                recipient_name: String,
-                kind_name: String,
-                payload: Vec<u8>,
-                count: u32,
-            }
-            #[derive(serde::Deserialize, Debug)]
-            struct CaptureFrameMirror {
-                mails: Vec<EnvelopeMirror>,
-                after_mails: Vec<EnvelopeMirror>,
-            }
-            let HubToEngine::Mail(frame) = req else {
-                panic!("expected Mail");
-            };
-            let decoded: CaptureFrameMirror =
-                postcard::from_bytes(&frame.payload).expect("decode CaptureFrame");
-            assert_eq!(decoded.mails.len(), 1);
-            assert_eq!(decoded.mails[0].recipient_name, "target.box");
-            assert_eq!(decoded.mails[0].kind_name, "demo.tick");
-            assert_eq!(decoded.mails[0].count, 1);
-            // `demo.tick` is Unit — payload is zero bytes.
-            assert!(decoded.mails[0].payload.is_empty());
-            // Test passes `after_mails: vec![]` — empty bundle.
-            assert!(decoded.after_mails.is_empty());
-
-            // Reply with a fake PNG so the tool completes.
-            #[derive(serde::Serialize)]
-            enum ReplyMirror {
-                #[allow(dead_code)]
-                Ok { png: Vec<u8> },
-                #[allow(dead_code)]
-                Err { error: String },
-            }
-            let reply = postcard::to_allocvec(&ReplyMirror::Ok {
-                png: vec![0x89, b'P', b'N', b'G', 0x0d, 0x0a, 0x1a, 0x0a, 42],
-            })
-            .expect("encode");
-            let record = sessions_clone.get(&session_token).expect("session");
-            let kind: String = "aether.render.capture_frame_result".into();
-            let queued = QueuedMail {
-                engine_id: id,
-                kind_name: kind.clone(),
-                payload: reply,
-                broadcast: false,
-                origin: None,
-            };
-            record.replies.try_deliver(&kind, queued);
-        });
-
-        let result = hub
-            .capture_frame(Parameters(CaptureFrameArgs {
-                engine_id: id.0.to_string(),
-                mails: vec![MailSpec {
-                    engine_id: id.0.to_string(),
-                    recipient_name: "target.box".into(),
-                    kind_name: "demo.tick".into(),
-                    params: None,
-                    count: 1,
-                }],
-                after_mails: vec![],
-                timeout_ms: Some(2_000),
-            }))
-            .await
-            .expect("bundle tool should succeed");
-
-        substrate_task.await.expect("stub substrate");
-        assert!(result.content[0].as_image().is_some());
-    }
-
-    #[tokio::test]
-    async fn capture_frame_rejects_bundle_with_unknown_kind() {
-        // Bundle contains a kind the engine doesn't declare. The tool
-        // should abort before sending anything and surface a clear
-        // invalid-params error naming the offending envelope.
-        let engines = EngineRegistry::new();
-        // Only capture_frame is declared; `demo.mystery` is not.
-        let kinds = vec![capture_frame_kind_descriptor()];
-        let (rec, _rx) = record_with_kinds(782, kinds);
-        let id = rec.id;
-        engines.insert(rec);
-        let state = test_state(engines, SessionRegistry::new());
-        let hub = Hub::new(state);
-
-        let err = hub
-            .capture_frame(Parameters(CaptureFrameArgs {
-                engine_id: id.0.to_string(),
-                mails: vec![MailSpec {
-                    engine_id: id.0.to_string(),
-                    recipient_name: "any".into(),
-                    kind_name: "demo.mystery".into(),
-                    params: None,
-                    count: 1,
-                }],
-                after_mails: vec![],
-                timeout_ms: Some(1_000),
-            }))
-            .await
-            .unwrap_err();
-        let msg = format!("{err:?}");
-        assert!(
-            msg.contains("no descriptor") && msg.contains("demo.mystery"),
-            "expected abort with kind name: {msg}"
-        );
-    }
-
-    // ADR-0078 Phase 1 (PR 596 + this PR): the spawn / terminate
-    // tool tests retired alongside the bespoke `spawn_substrate` /
-    // `terminate_substrate` helpers — both now route through
-    // `ProcessCapability`. The cap's own unit tests in
-    // `crate::hub::process_capability` cover spawn-handshake-timeout
-    // (`spawn_handshake_timeout_replies_err`), reaper broadcast
-    // (`reaper_emits_process_exited_when_child_exits`), and cap
-    // boot. End-to-end MCP→cap coverage rides on the smoke flow
-    // documented in the PR body.
-}
