--- crates/aether-substrate-bundle/src/hub/mcp/args.rs
@@ -1,602 +1,0 @@
-//! Request / response shapes for the MCP tool surface. These types
-//! only carry data — no hub state, no business logic. Each request type
-//! is deserialized from the MCP tool call's `Parameters`, and each
-//! response type is serialized back out as JSON. `JsonSchema` is
-//! required on every type that appears in a tool signature so rmcp
-//! can publish the correct schema to the client.
-
-use std::collections::HashMap;
-
-use crate::hub::wire::LogLevel;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct DescribeKindsArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct DescribeComponentArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Mailbox id of the loaded component (from `load_component`'s
-    /// response). ADR-0064 tagged-string form: `mbx-XXXX-XXXX-XXXX`,
-    /// 12 lowercase base32 chars (RFC 4648 alphabet) grouped 4-4-4
-    /// after the `mbx-` prefix. Strings, not numbers, because
-    /// JSON's IEEE-754 double rounds 64-bit ids past 2^53; the
-    /// prefix also makes copy-paste mistakes between mailbox / kind /
-    /// handle ids reject at the boundary instead of failing later.
-    pub mailbox_id: String,
-}
-
-/// Structured response for `describe_component` (ADR-0033). Mirrors
-/// the `LoadResult` capabilities but adds the component's name so an
-/// agent that only has the mailbox id can still render it.
-#[derive(Debug, Serialize, JsonSchema)]
-pub struct DescribeComponentResponse {
-    pub name: String,
-    pub doc: Option<String>,
-    pub receives: Vec<HandlerCapabilityWire>,
-    pub fallback: Option<FallbackCapabilityWire>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct SendMailArgs {
-    /// One or more mail items to deliver. Each item is processed
-    /// independently — a single failure doesn't abort the batch. The
-    /// response carries a per-item status so the caller decides retry
-    /// vs abort policy.
-    pub mails: Vec<MailSpec>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct MailSpec {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Mailbox name as registered by the engine.
-    pub recipient_name: String,
-    /// Kind name (e.g. `"aether.tick"`) the engine's registry knows.
-    pub kind_name: String,
-    /// Structured params for schema-driven encoding. The hub looks up
-    /// this kind's descriptor (from `describe_kinds`) and writes bytes
-    /// matching the kind's wire format — `#[repr(C)]` layout for
-    /// cast-shaped kinds, postcard for everything else.
-    ///
-    /// ADR-0019 PR 5 removed the `payload_bytes` escape hatch from
-    /// this surface. Every aether-shipped kind has a schema; if a
-    /// future kind doesn't, that's an engine bug to fix, not a
-    /// workaround to paper over.
-    #[serde(default)]
-    pub params: Option<serde_json::Value>,
-    /// Count carried on the mail frame. For single-struct payloads
-    /// this is 1 (the default); for cast-shaped slices it's the
-    /// number of elements the bytes represent.
-    #[serde(default = "one")]
-    pub count: u32,
-}
-
-#[derive(Debug, Serialize, Deserialize, JsonSchema)]
-pub struct MailStatus {
-    /// Index into the `mails` array the caller supplied.
-    pub index: u32,
-    /// `"delivered"` on success, or `"error: <reason>"` on failure.
-    pub status: String,
-}
-
-pub(super) fn one() -> u32 {
-    1
-}
-
-pub(super) fn log_level_to_str(level: LogLevel) -> &'static str {
-    match level {
-        LogLevel::Trace => "trace",
-        LogLevel::Debug => "debug",
-        LogLevel::Info => "info",
-        LogLevel::Warn => "warn",
-        LogLevel::Error => "error",
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, JsonSchema)]
-pub struct EngineInfo {
-    pub engine_id: String,
-    pub name: String,
-    pub pid: u32,
-    pub version: String,
-    /// `true` if this engine was spawned by the hub (ADR-0009), `false`
-    /// if it connected externally.
-    pub spawned: bool,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct SpawnSubstrateArgs {
-    /// Absolute path to the substrate binary the hub should launch.
-    /// The hub does not build, resolve, or locate binaries — the caller
-    /// passes a path that exists.
-    pub binary_path: String,
-    /// Additional command-line arguments to pass to the substrate.
-    #[serde(default)]
-    pub args: Vec<String>,
-    /// Extra environment variables for the child. `AETHER_HUB_URL` is
-    /// injected automatically to point at this hub; if the caller
-    /// includes it here, the caller's value wins.
-    #[serde(default)]
-    pub env: HashMap<String, String>,
-    /// Handshake timeout in milliseconds. Defaults to 5 seconds if
-    /// omitted — override for slow CI machines or debug builds.
-    /// Applies both to the substrate's `Hello` handshake and to each
-    /// component load in `components` (each load waits up to this
-    /// timeout for its `LoadResult` reply).
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-    /// Optional component preloads. After the substrate finishes its
-    /// `Hello` handshake, the hub reads each component's wasm and
-    /// loads it via the same path `load_component` uses, sequentially
-    /// in the declared order. All-or-nothing: a single load failure
-    /// SIGTERMs the freshly-spawned child and bubbles the error — no
-    /// half-loaded scene is ever returned to the caller.
-    ///
-    /// Ordering matters for scenes where one component mails another
-    /// by name at init time (e.g. a world component that mails a
-    /// camera by its loaded name on first tick).
-    #[serde(default)]
-    pub components: Vec<SpawnComponentSpec>,
-}
-
-/// One component to preload during `spawn_substrate`. Shape matches the
-/// relevant subset of `LoadComponentArgs`: a filesystem path and an
-/// optional caller-chosen name. No `timeout_ms` here — the spawn-level
-/// timeout governs each load.
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct SpawnComponentSpec {
-    /// Absolute path to the WASM component binary on the hub's
-    /// filesystem. Same path rule as top-level `binary_path`: no `~`
-    /// expansion, no relative resolution.
-    pub binary_path: String,
-    /// Optional human-readable name, passed through to the substrate's
-    /// `LoadComponent`. The substrate defaults one when omitted.
-    #[serde(default)]
-    pub name: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, JsonSchema)]
-pub struct SpawnResult {
-    /// UUID assigned by the hub once the substrate completed its
-    /// `Hello` handshake.
-    pub engine_id: String,
-    /// Operating-system PID of the spawned substrate.
-    pub pid: u32,
-    /// Per-component results for anything listed in the request's
-    /// `components` array, in the same order. Empty when no preloads
-    /// were requested. Every entry is a success; an all-or-nothing
-    /// failure surfaces as a tool-level error instead of a partial
-    /// `SpawnResult`.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub components: Vec<LoadComponentResponse>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct TerminateSubstrateArgs {
-    /// Hub-assigned engine UUID (from `list_engines`).
-    pub engine_id: String,
-    /// SIGTERM grace period in milliseconds. Defaults to 2 seconds —
-    /// long enough for a well-behaved substrate to drain, short enough
-    /// not to stall interactive agent flows.
-    #[serde(default)]
-    pub grace_ms: Option<u32>,
-}
-
-#[derive(Debug, Serialize, Deserialize, JsonSchema)]
-pub struct TerminateResult {
-    pub engine_id: String,
-    /// `true` if the child ignored SIGTERM and the hub escalated to
-    /// SIGKILL after the grace window. `false` for clean exits.
-    pub sigkilled: bool,
-    /// Exit code if the child exited normally; `null` if it was killed
-    /// by a signal (the common case when `sigkilled` is true).
-    pub exit_code: Option<i32>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct EngineLogsArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Maximum entries to return. Defaults to 100; clamped to 1000.
-    #[serde(default)]
-    pub max: Option<u32>,
-    /// Minimum log level (`"trace"|"debug"|"info"|"warn"|"error"`).
-    /// Defaults to `"trace"` (every captured entry).
-    #[serde(default)]
-    pub level: Option<String>,
-    /// Cursor: only entries with `sequence > since` are returned.
-    /// Pass back the previous response's `next_since` to poll
-    /// incrementally without re-receiving.
-    #[serde(default)]
-    pub since: Option<u64>,
-}
-
-#[derive(Debug, Serialize, JsonSchema)]
-pub struct EngineLogsResponse {
-    pub engine_id: String,
-    pub entries: Vec<EngineLogEntry>,
-    /// Highest `sequence` returned in `entries`, or the request's
-    /// `since` value if the response is empty. Use as the next
-    /// `since` for incremental polling.
-    pub next_since: u64,
-    /// Set when the hub-side ring evicted entries the caller hadn't
-    /// seen — `Some(seq)` means the gap above the request's `since`
-    /// extends up to (but not including) `seq`. Treat as a signal to
-    /// poll more often or accept the loss.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub truncated_before: Option<u64>,
-}
-
-#[derive(Debug, Serialize, JsonSchema)]
-pub struct EngineLogEntry {
-    pub timestamp_unix_ms: u64,
-    /// Lowercase level: `"trace"|"debug"|"info"|"warn"|"error"`.
-    pub level: String,
-    /// Module path the event was emitted from
-    /// (e.g. `"aether_substrate::scheduler"`).
-    pub target: String,
-    /// Already-formatted event text. Tracing's structured fields are
-    /// flattened into this string by the substrate's capture layer.
-    pub message: String,
-    pub sequence: u64,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct CaptureFrameArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Optional bundle of mails to dispatch on the substrate *before*
-    /// the capture fires. Each item has the same shape as a
-    /// `send_mail` entry — `recipient_name`, `kind_name`, structured
-    /// `params` encoded via the kind's descriptor, `count`. The
-    /// substrate resolves every envelope atomically: if any fails
-    /// (unknown kind, unknown recipient), no mail is dispatched and
-    /// the reply is an `Err`. An empty bundle means "just capture the
-    /// current state."
-    #[serde(default)]
-    pub mails: Vec<MailSpec>,
-    /// Optional bundle of cleanup mails dispatched *after* the frame
-    /// is captured. Useful for "set a flag, capture, unset the flag"
-    /// patterns in one atomic tool call. Resolved against the same
-    /// abort-on-first-failure policy as `mails` — a bad envelope in
-    /// either bundle aborts the whole request before any mail moves.
-    #[serde(default)]
-    pub after_mails: Vec<MailSpec>,
-    /// Maximum time to wait for the substrate's reply, in
-    /// milliseconds. Defaults to 5000 (5 s). Clamped to 30000 (30 s).
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-}
-
-/// Wire-format mirror of the substrate's `CaptureFrameResult` kind.
-/// Lives in the hub so we can postcard-decode the reply payload
-/// without pulling in the substrate's typed kind crate. Must stay in
-/// lockstep with `aether-kinds::CaptureFrameResult` — the comment on
-/// that type is the canonical spec.
-#[derive(Debug, Deserialize)]
-pub(super) enum CaptureFrameResultWire {
-    Ok { png: Vec<u8> },
-    Err { error: String },
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct LoadComponentArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Absolute path to the WASM component binary on the hub's
-    /// filesystem. The hub reads the file and sends the bytes to the
-    /// substrate — agents don't inline wasm bytes through the tool
-    /// call. Must exist as given; no `~` expansion, no relative-path
-    /// resolution. Matches `spawn_substrate`'s path rule.
-    ///
-    /// ADR-0028: the component's kind vocabulary is embedded in the
-    /// wasm's `aether.kinds` custom section at compile time. The
-    /// loader doesn't declare kinds — the substrate reads them
-    /// straight from the binary.
-    pub binary_path: String,
-    /// Optional human-readable name. The substrate defaults one if
-    /// absent and echoes it back in the result.
-    #[serde(default)]
-    pub name: Option<String>,
-    /// Maximum time to wait for the substrate's `LoadResult` reply,
-    /// in milliseconds. Defaults to 5000. Clamped to 30000.
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-}
-
-/// Wire-format mirror of `aether.component.load_result`. Postcard-
-/// decoded from the substrate's reply payload. Must stay in lockstep
-/// with `aether-kinds::LoadResult` — that type is canonical.
-#[derive(Debug, Deserialize)]
-pub(super) enum LoadResultWire {
-    Ok {
-        mailbox_id: u64,
-        name: String,
-        capabilities: ComponentCapabilitiesWire,
-    },
-    Err {
-        error: String,
-    },
-}
-
-/// Wire-format mirror of `aether-kinds::ComponentCapabilities` (ADR-
-/// 0033). The hub decodes this from the substrate's `LoadResult` and
-/// caches a per-mailbox copy so the `describe_component` MCP tool can
-/// surface it without round-tripping the substrate. Same lockstep
-/// discipline as `LoadResultWire` — canonical form lives in aether-kinds.
-#[derive(Clone, Debug, Default, Deserialize, Serialize, JsonSchema)]
-pub struct ComponentCapabilitiesWire {
-    pub handlers: Vec<HandlerCapabilityWire>,
-    pub fallback: Option<FallbackCapabilityWire>,
-    pub doc: Option<String>,
-}
-
-/// One `#[handler]` method's capability entry — see `aether-kinds::HandlerCapability`.
-///
-/// `id` is the typed `KindId`. ADR-0064 / ADR-0065: `KindId`'s serde
-/// impl emits the tagged-string form (`knd-XXXX-XXXX-XXXX`) for
-/// human-readable serializers (JSON) and the raw `u64` for binary
-/// (postcard), so the same field deserializes from the substrate's
-/// postcard frame and serializes out to the MCP JSON response without
-/// any manual conversion. The `#[schemars(with = "String")]` attribute
-/// is required because `aether-data` is `no_std + alloc` and doesn't
-/// depend on schemars; the attribute publishes a `String` schema for
-/// the JSON wire form while leaving serde untouched.
-#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
-pub struct HandlerCapabilityWire {
-    #[schemars(with = "String")]
-    pub id: aether_data::KindId,
-    pub name: String,
-    pub doc: Option<String>,
-}
-
-/// `#[fallback]` presence + optional doc — see `aether-kinds::FallbackCapability`.
-#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
-pub struct FallbackCapabilityWire {
-    pub doc: Option<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, JsonSchema)]
-pub struct LoadComponentResponse {
-    /// Substrate-assigned mailbox id for the loaded component, in the
-    /// ADR-0064 tagged-string form `mbx-XXXX-XXXX-XXXX`. Hand this to
-    /// `replace_component` or `describe_component` verbatim.
-    pub mailbox_id: String,
-    /// Substrate-resolved name. Matches the `name` in the request if
-    /// provided; otherwise the substrate-defaulted value.
-    pub name: String,
-    /// Receive-side capability surface the component advertised via
-    /// `aether.kinds.inputs` (ADR-0033). Empty `handlers` + no
-    /// fallback + no doc describes a component that shipped without
-    /// the `#[actor]` macro — still loadable, just opaque to
-    /// `describe_component`.
-    pub capabilities: ComponentCapabilitiesWire,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct ReplaceComponentArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Mailbox id of the live component to replace, in the ADR-0064
-    /// tagged-string form `mbx-XXXX-XXXX-XXXX` (from a prior
-    /// `load_component` or `list_engines`-derived lookup).
-    pub mailbox_id: String,
-    /// Absolute path to the replacement WASM binary on the hub's
-    /// filesystem. Same filesystem rule as `load_component`. Kind
-    /// vocabulary is embedded in the wasm's `aether.kinds` custom
-    /// section (ADR-0028).
-    pub binary_path: String,
-    /// Drain timeout in milliseconds for in-flight mail on the old
-    /// instance (ADR-0022). `None` uses the substrate default (5000).
-    /// If the drain exceeds this, the replace fails and the old
-    /// instance stays bound.
-    #[serde(default)]
-    pub drain_timeout_ms: Option<u32>,
-    /// Maximum time to wait for the substrate's `ReplaceResult` reply,
-    /// in milliseconds. Defaults to 5000. Clamped to 30000. Set higher
-    /// than `drain_timeout_ms` so the reply has room to arrive after
-    /// the substrate declares drain failure.
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-}
-
-/// Wire-format mirror of `aether.component.replace_result`. Must stay
-/// in lockstep with `aether-kinds::ReplaceResult`.
-#[derive(Debug, Deserialize)]
-pub(super) enum ReplaceResultWire {
-    Ok {
-        capabilities: ComponentCapabilitiesWire,
-    },
-    Err {
-        error: String,
-    },
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct ReceiveMailArgs {
-    /// Maximum number of items to return in this call. `None` drains
-    /// everything currently queued. Defaults to unlimited.
-    #[serde(default)]
-    pub max: Option<u32>,
-}
-
-#[derive(Debug, Serialize, Deserialize, JsonSchema)]
-pub struct ReceivedMail {
-    /// Hub-assigned UUID of the engine that sent this mail.
-    pub engine_id: String,
-    /// Kind name the engine declared at handshake.
-    pub kind_name: String,
-    /// Structured value decoded against the engine's kind descriptor —
-    /// the symmetric counterpart to `send_mail`'s `params` field
-    /// (ADR-0020). `null` if the hub couldn't decode (no descriptor for
-    /// this kind, or decode failed), in which case `decode_error` is
-    /// populated and the agent falls back to `payload_bytes`.
-    pub params: Option<serde_json::Value>,
-    /// Decode failure reason. Populated only when `params` is `null`
-    /// because the hub's decoder rejected the payload (e.g. schema
-    /// drift, truncation, unknown kind). Always paired with intact
-    /// `payload_bytes` for an escape-hatch decode.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub decode_error: Option<String>,
-    /// Raw payload bytes. Always populated. Prefer `params`; reach for
-    /// `payload_bytes` only when `params` is `null` (decode failed) or
-    /// when the agent genuinely needs the wire bytes.
-    pub payload_bytes: Vec<u8>,
-    /// `true` if this mail was addressed to every attached session;
-    /// `false` if it was a reply targeted at this session specifically.
-    pub broadcast: bool,
-    /// Substrate-attested name of the emitting mailbox (ADR-0011).
-    /// Distinguishes components that share a kind. `None` for mail
-    /// pushed by substrate core (e.g. the frame loop's `FrameStats`),
-    /// which has no sending mailbox.
-    pub origin: Option<String>,
-}
-
-/// Issue 718: agent-facing JSON shape of `aether_data::MailId`. The
-/// underlying type is a 128-bit composite (`MailboxId` sender + `u64`
-/// correlation) — too wide for the ADR-0064 single-string tagged-id
-/// scheme — so the wire form mirrors the struct, with `sender` carrying
-/// the existing `mbx-XXXX-XXXX-XXXX` tagged form. `correlation_id`
-/// stays a `u64`; for short-running processes it's well under 2^53 and
-/// the agent treats it as opaque (round-trip without arithmetic).
-#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
-pub struct MailIdWire {
-    /// Tagged-string mailbox id of the sender that minted the mail.
-    pub sender: String,
-    /// Per-actor correlation counter; opaque to the agent.
-    pub correlation_id: u64,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct ListActiveRootsArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Filter to roots whose originating `Sent` event was emitted no
-    /// more than this many milliseconds ago. Defaults to 60_000 (1
-    /// minute). Roots older than the window aren't returned even if
-    /// they haven't been evicted yet.
-    #[serde(default)]
-    pub since_ms: Option<u32>,
-    /// Cap the reply length. Defaults to 50; clamped to 1000 by the
-    /// observer. Sorted desc by `t_sent` so the most recent roots
-    /// always land in the truncated head.
-    #[serde(default)]
-    pub max: Option<u32>,
-    /// Maximum time to wait for the substrate's reply, in
-    /// milliseconds. Defaults to 5000. Clamped to 30000.
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct DumpTraceArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Root mail id whose subtree should be exported. Same shape as
-    /// `describe_tree`'s `root` argument — a `MailNodeWire.root` from
-    /// a prior `describe_tree` reply or a `RootSummaryWire.root` from
-    /// `list_active_roots`.
-    pub root: MailIdWire,
-    /// Optional filesystem path to write the trace JSON to. Parent
-    /// directories are created as needed. When omitted, the JSON is
-    /// returned inline as the tool result (fine for small subtrees;
-    /// for busy substrates that generate thousands of events the
-    /// inline path may overflow the agent's context).
-    #[serde(default)]
-    pub output_path: Option<String>,
-    /// Maximum time to wait for the substrate's `DescribeTreeResult`
-    /// reply, in milliseconds. Defaults to 5000. Clamped to 30000.
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-}
-
-/// Issue 735: agent-facing window selector for the time-window trace
-/// queries. Untagged so the JSON discriminates by field presence —
-/// `{"last_ms": N}` is `Relative`, `{"start_ns": X, "end_ns": Y?}` is
-/// `Absolute`. The two field-name sets don't overlap, so deserialize
-/// is unambiguous. Mirrors `aether_kinds::trace::TraceWindow`; kept
-/// hub-side because `aether-kinds` is `no_std + alloc` and doesn't
-/// depend on `schemars`.
-#[derive(Debug, Deserialize, JsonSchema)]
-#[serde(untagged)]
-pub enum TraceWindowArg {
-    /// Explicit absolute range in nanoseconds-since-substrate-boot.
-    /// `end_ns: None` means "open-ended through now" (resolved by the
-    /// observer at handler entry).
-    Absolute {
-        start_ns: u64,
-        #[serde(default)]
-        end_ns: Option<u64>,
-    },
-    /// Last N milliseconds, relative to the substrate's monotonic now
-    /// at handler entry. Equivalent to `{ start_ns: now - last_ms, end_ns: None }`.
-    Relative { last_ms: u64 },
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct DescribeTreeWindowArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Time window selector. Either `{"last_ms": N}` for "last N
-    /// milliseconds" (resolved against substrate monotonic now) or
-    /// `{"start_ns": ..., "end_ns": ...}` for an explicit absolute
-    /// range. `end_ns` is optional; omitting it means "open-ended
-    /// through now."
-    pub window: TraceWindowArg,
-    /// Cap on the number of in-window mails returned. Substrate
-    /// defaults 10_000; hard cap 100_000. Over-cap windows reply
-    /// `Err::too_many` with the matched count so you can narrow the
-    /// window or raise this. Strict containment of `t_sent`.
-    #[serde(default)]
-    pub max_mails: Option<u32>,
-    /// Maximum time to wait for the substrate's reply, in
-    /// milliseconds. Defaults to 5000. Clamped to 30000.
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct DumpTraceWindowArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Time window selector. Same shape as `describe_tree_window`'s
-    /// `window` arg — `{"last_ms": N}` for relative or
-    /// `{"start_ns": ..., "end_ns": ...}` for absolute. `end_ns`
-    /// optional.
-    pub window: TraceWindowArg,
-    /// Optional filesystem path to write the trace JSON to. Parent
-    /// directories are created as needed. When omitted the JSON is
-    /// returned inline; for any non-trivial window that's likely
-    /// to overflow the agent's context — point this at a file
-    /// instead and load it into Perfetto.
-    #[serde(default)]
-    pub output_path: Option<String>,
-    /// Cap on the number of in-window mails the substrate will return
-    /// before erroring with `Err::too_many`. Substrate defaults
-    /// 10_000; hard cap 100_000. Mirrors `describe_tree_window`.
-    #[serde(default)]
-    pub max_mails: Option<u32>,
-    /// Maximum time to wait for the substrate's `DescribeWindowResult`
-    /// reply, in milliseconds. Defaults to 5000. Clamped to 30000.
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-}
-
-#[derive(Debug, Deserialize, JsonSchema)]
-pub struct DescribeTreeArgs {
-    /// Hub-assigned engine UUID as a string (from `list_engines`).
-    pub engine_id: String,
-    /// Root mail id to describe. Obtained from a prior
-    /// `list_active_roots` reply or from any observed `MailNodeWire.root`.
-    pub root: MailIdWire,
-    /// Maximum time to wait for the substrate's reply, in
-    /// milliseconds. Defaults to 5000. Clamped to 30000.
-    #[serde(default)]
-    pub timeout_ms: Option<u32>,
-}
