--- crates/aether-substrate-bundle/src/hub/mcp/tools.rs
@@ -1,1338 +1,0 @@
-//! The `#[tool_router] impl Hub` block — every MCP tool the hub
-//! exposes lives here. Each method is a thin adapter that parses
-//! arguments, looks up registry state, delegates to the helpers in
-//! `codecs.rs` for encode/decode work, and serializes the response.
-//! Business logic that isn't request/response glue belongs in
-//! `crate::hub::registry`, `crate::hub::process_capability`,
-//! `crate::hub::session`, or the codec module — not here.
-
-use std::sync::Arc;
-use std::time::Duration;
-
-use crate::hub::wire::{EngineId, LogLevel, Uuid};
-use aether_actor::Actor;
-use aether_data::Kind;
-use aether_data::tagged_id::{self, Tag};
-use aether_kinds::{
-    EnvVar, Spawn as WireSpawn, SpawnResult as WireSpawnResult, Terminate as WireTerminate,
-    TerminateResult as WireTerminateResult,
-    trace::{
-        DescribeTree, DescribeTreeResult, DescribeWindow, DescribeWindowResult, ListActiveRoots,
-        ListActiveRootsResult, TRACE_OBSERVER_MAILBOX_NAME, TraceWindow,
-    },
-};
-use base64::Engine as _;
-use rmcp::{
-    ErrorData as McpError, ServerHandler,
-    handler::server::wrapper::Parameters,
-    model::{CallToolResult, Content, Implementation, ServerCapabilities, ServerInfo},
-    tool, tool_handler, tool_router,
-};
-use tokio::sync::Mutex;
-use tokio::time::timeout;
-
-use crate::hub::log_store::{TOOL_DEFAULT_ENTRIES, TOOL_MAX_ENTRIES};
-use crate::hub::loopback::HUB_SELF_ENGINE_ID;
-use crate::hub::process_capability::ProcessCapability;
-use crate::hub::session::SessionHandle;
-
-use super::args::{
-    CaptureFrameArgs, CaptureFrameResultWire, DescribeComponentArgs, DescribeComponentResponse,
-    DescribeKindsArgs, DescribeTreeArgs, DescribeTreeWindowArgs, DumpTraceArgs,
-    DumpTraceWindowArgs, EngineInfo, EngineLogEntry, EngineLogsArgs, EngineLogsResponse,
-    ListActiveRootsArgs, LoadComponentArgs, LoadComponentResponse, LoadResultWire, MailSpec,
-    MailStatus, ReceiveMailArgs, ReceivedMail, ReplaceComponentArgs, ReplaceResultWire,
-    SendMailArgs, SpawnResult, SpawnSubstrateArgs, TerminateResult, TerminateSubstrateArgs,
-    TraceWindowArg, log_level_to_str,
-};
-use super::codecs::{decode_inbound, deliver_one, encode_capture_bundle};
-use super::trace::{render, render_mails};
-use super::{Hub, HubState};
-use crate::hub::registry::ComponentRecord;
-
-/// Substrate control-plane kind names. Predate the `aether-kinds`
-/// dep this crate now carries (PR 596 introduced it for the process
-/// kinds); future cleanup migrates these to `<K as Kind>::NAME` at
-/// the call site. The process tools below already use that pattern;
-/// these survive to avoid expanding PR 2's diff into the existing
-/// capture / load / replace paths.
-const KIND_CAPTURE_FRAME: &str = "aether.render.capture_frame";
-const KIND_CAPTURE_FRAME_RESULT: &str = "aether.render.capture_frame_result";
-const KIND_LOAD_COMPONENT: &str = "aether.component.load";
-const KIND_LOAD_RESULT: &str = "aether.component.load_result";
-const KIND_REPLACE_COMPONENT: &str = "aether.component.replace";
-const KIND_REPLACE_RESULT: &str = "aether.component.replace_result";
-
-/// Shared default/cap for await-reply tools. `DEFAULT_CAPTURE_TIMEOUT`
-/// is reused directly by name from the capture path; these aliases
-/// exist so the load/replace paths don't have to reach across the
-/// implementation to share the constant. Same values: 5s default, 30s
-/// ceiling.
-const DEFAULT_REPLY_TIMEOUT: Duration = DEFAULT_CAPTURE_TIMEOUT;
-const MAX_REPLY_TIMEOUT: Duration = MAX_CAPTURE_TIMEOUT;
-
-/// Default cap on how long `capture_frame` waits for the substrate's
-/// reply before returning an error. Long enough to tolerate one
-/// stalled frame (monitor refresh + GPU readback + PNG encode); short
-/// enough that a stuck substrate doesn't hang the tool call.
-const DEFAULT_CAPTURE_TIMEOUT: Duration = Duration::from_secs(5);
-
-/// Hard ceiling on `timeout_ms`. A request that needs more than 30s
-/// is probably a bug on the substrate side; surfacing it as a bound
-/// rather than letting it hang preserves the harness's responsiveness.
-const MAX_CAPTURE_TIMEOUT: Duration = Duration::from_secs(30);
-
-/// Internal mirror of the wire `aether.process.terminate_result::Ok`
-/// variant. Returned by `Hub::do_terminate_via_cap` to both the
-/// public `terminate_substrate` MCP tool and the spawn-substrate
-/// preload-cleanup path.
-struct TerminateOutcomeWire {
-    exit_code: Option<i32>,
-    sigkilled: bool,
-}
-
-impl Hub {
-    /// Construct a per-session `Hub`. Lives in this module (rather
-    /// than `mcp.rs` with the struct) because `Self::tool_router()` is
-    /// generated private by rmcp's `#[tool_router]` macro below and
-    /// only siblings in the same module can call it.
-    pub fn new(state: Arc<HubState>) -> Self {
-        let (session, rx) = SessionHandle::mint(&state.sessions);
-        Self {
-            state,
-            tool_router: Self::tool_router(),
-            session: Arc::new(session),
-            inbound: Arc::new(Mutex::new(rx)),
-        }
-    }
-}
-
-#[tool_router]
-impl Hub {
-    #[tool(
-        description = "Send one or more mail items. Each item takes `params` — structured JSON encoded via the engine's kind descriptor (cast-shaped or postcard, ADR-0019). The batch is best-effort: per-item status is returned and failures don't abort siblings. 'delivered' means the hub queued the frame to the engine's socket — not that the engine processed it."
-    )]
-    pub(super) async fn send_mail(
-        &self,
-        Parameters(args): Parameters<SendMailArgs>,
-    ) -> Result<String, McpError> {
-        let mut statuses = Vec::with_capacity(args.mails.len());
-        for (i, spec) in args.mails.into_iter().enumerate() {
-            let result = deliver_one(spec, &self.state.engines, self.session.token).await;
-            let status = match result {
-                Ok(()) => "delivered".into(),
-                Err(e) => format!("error: {e}"),
-            };
-            statuses.push(MailStatus {
-                index: i as u32,
-                status,
-            });
-        }
-        serde_json::to_string(&statuses).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    #[tool(
-        description = "List every kind the given engine declared at handshake, with enough structural detail for clients to build params for send_mail. ADR-0019 Schema kinds describe their full shape (scalars, strings, vecs, options, enums, nested structs); the cast-shaped subset (`Struct{repr_c:true}`) is wire-compatible with `#[repr(C)]` layout."
-    )]
-    pub(super) async fn describe_kinds(
-        &self,
-        Parameters(args): Parameters<DescribeKindsArgs>,
-    ) -> Result<String, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        let Some(record) = self.state.engines.get(&id) else {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        };
-        serde_json::to_string(&record.kinds)
-            .map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    #[tool(
-        description = "Drain observation mail addressed to this MCP session. Returns everything currently queued (up to `max`, if provided). Each item reports the originating engine_id, the kind name, structured `params` decoded against the engine's kind descriptor (ADR-0020 — symmetric to `send_mail`), the raw `payload_bytes` (always populated; primarily a fallback when `params` is null), an optional `decode_error` explaining why decode failed when `params` is null, a `broadcast` flag indicating whether this mail also went to every other attached session (true) or was targeted specifically at this one (false), and an optional `origin` — the substrate-local mailbox name of the emitting component (absent for substrate-core pushes with no sending mailbox). Non-blocking: returns an empty array if nothing is queued."
-    )]
-    pub(super) async fn receive_mail(
-        &self,
-        Parameters(args): Parameters<ReceiveMailArgs>,
-    ) -> Result<String, McpError> {
-        let cap = args.max.map_or(usize::MAX, |n| n as usize);
-        let mut rx = self.inbound.lock().await;
-        let mut out = Vec::new();
-        while out.len() < cap {
-            match rx.try_recv() {
-                Ok(m) => {
-                    let (params, decode_error) =
-                        decode_inbound(&m.engine_id, &m.kind_name, &m.payload, &self.state.engines);
-                    out.push(ReceivedMail {
-                        engine_id: m.engine_id.0.to_string(),
-                        kind_name: m.kind_name,
-                        params,
-                        decode_error,
-                        payload_bytes: m.payload,
-                        broadcast: m.broadcast,
-                        origin: m.origin,
-                    });
-                }
-                Err(_) => break,
-            }
-        }
-        serde_json::to_string(&out).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    #[tool(
-        description = "Launch a substrate binary as a child process of the hub. The hub injects `AETHER_HUB_URL` so the child dials back to this hub instance. Blocks until the substrate completes its `Hello` handshake (or the handshake timeout fires — default 5 seconds, overridable via `timeout_ms`). Returns the assigned `engine_id`, the child `pid`, and per-component results for anything in the optional `components` preload list. Preloads are all-or-nothing: any load failure SIGTERMs the freshly-spawned child and surfaces as a tool error. The hub owns the child for its lifetime; dropping the engine from the registry (socket disconnect or `terminate_substrate`) reaps the process."
-    )]
-    pub(super) async fn spawn_substrate(
-        &self,
-        Parameters(args): Parameters<SpawnSubstrateArgs>,
-    ) -> Result<String, McpError> {
-        // ADR-0078 Phase 1: route through `ProcessCapability` rather
-        // than calling `crate::hub::spawn::spawn_substrate` directly.
-        // The cap owns the freshly-spawned `Child` and the per-child
-        // reaper task; the MCP coordinator becomes a thin wrapper
-        // that mails the cap and awaits its reply via the same
-        // session-reply diversion `capture_frame` / `load_component`
-        // use.
-        let env: Vec<EnvVar> = args
-            .env
-            .into_iter()
-            .map(|(key, value)| EnvVar { key, value })
-            .collect();
-        let params = serde_json::json!({
-            "binary_path": args.binary_path,
-            "args": args.args,
-            "env": env,
-            "handshake_timeout_ms": args.timeout_ms,
-        });
-
-        let (_guard, rx) = self
-            .session
-            .replies
-            .register(WireSpawnResult::NAME.to_owned())
-            .ok_or_else(|| {
-                McpError::invalid_params(
-                    "a spawn_substrate is already in flight on this session; \
-                     wait for it to complete"
-                        .to_owned(),
-                    None,
-                )
-            })?;
-
-        let spec = MailSpec {
-            engine_id: HUB_SELF_ENGINE_ID.0.to_string(),
-            recipient_name: ProcessCapability::NAMESPACE.to_owned(),
-            kind_name: WireSpawn::NAME.to_owned(),
-            params: Some(params),
-            count: 1,
-        };
-        deliver_one(spec, &self.state.engines, self.session.token)
-            .await
-            .map_err(|e| McpError::internal_error(format!("send failed: {e}"), None))?;
-
-        let wait = args
-            .timeout_ms
-            .map(|ms| Duration::from_millis(ms as u64).min(MAX_REPLY_TIMEOUT))
-            .unwrap_or(DEFAULT_REPLY_TIMEOUT);
-        let queued = match timeout(wait, rx).await {
-            Ok(Ok(m)) => m,
-            Ok(Err(_)) => {
-                return Err(McpError::internal_error(
-                    "spawn_result channel closed before reply arrived".to_owned(),
-                    None,
-                ));
-            }
-            Err(_) => {
-                return Err(McpError::internal_error(
-                    format!(
-                        "timed out after {}ms waiting for spawn_result",
-                        wait.as_millis()
-                    ),
-                    None,
-                ));
-            }
-        };
-
-        let (engine_id_str, pid) = match postcard::from_bytes::<WireSpawnResult>(&queued.payload)
-            .map_err(|e| McpError::internal_error(format!("decode reply: {e}"), None))?
-        {
-            WireSpawnResult::Ok { engine_id, pid } => (engine_id, pid),
-            WireSpawnResult::Err { error } => {
-                return Err(McpError::internal_error(
-                    format!("spawn failed: {error}"),
-                    None,
-                ));
-            }
-        };
-        let engine_uuid = Uuid::parse_str(&engine_id_str).map_err(|e| {
-            McpError::internal_error(format!("cap returned non-UUID engine_id: {e}"), None)
-        })?;
-        let engine_id = EngineId(engine_uuid);
-
-        // Preload components sequentially. Any failure aborts the whole
-        // spawn: we mail `Terminate` to the cap before bubbling the
-        // error so a half-loaded substrate never leaks back to the
-        // caller as a successful-looking `SpawnResult`.
-        let mut components = Vec::with_capacity(args.components.len());
-        for (idx, spec) in args.components.into_iter().enumerate() {
-            match self
-                .do_load_component(
-                    &engine_id,
-                    &engine_id_str,
-                    &spec.binary_path,
-                    spec.name,
-                    args.timeout_ms,
-                )
-                .await
-            {
-                Ok(resp) => components.push(resp),
-                Err(load_err) => {
-                    let _ = self.do_terminate_via_cap(&engine_id_str, None, None).await;
-                    return Err(McpError::internal_error(
-                        format!(
-                            "preload #{idx} failed; spawned child terminated: {}",
-                            load_err.message
-                        ),
-                        None,
-                    ));
-                }
-            }
-        }
-
-        let result = SpawnResult {
-            engine_id: engine_id_str,
-            pid,
-            components,
-        };
-        serde_json::to_string(&result).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    #[tool(
-        description = "Terminate a substrate the hub previously spawned. Sends SIGTERM, waits up to `grace_ms` milliseconds (default 2000) for the child to exit, then escalates to SIGKILL if it's still running. Returns the exit code (if any) and a `sigkilled` flag indicating whether escalation was necessary. Errors if the engine id is unknown or refers to an externally connected substrate — the hub only terminates children it owns."
-    )]
-    pub(super) async fn terminate_substrate(
-        &self,
-        Parameters(args): Parameters<TerminateSubstrateArgs>,
-    ) -> Result<String, McpError> {
-        // ADR-0078 Phase 1: route through `ProcessCapability`. The cap
-        // owns the `Child`; the bespoke `crate::hub::spawn::terminate_substrate`
-        // is reserved for the chassis shutdown path and not addressable
-        // from MCP anymore.
-        let _ = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-
-        let outcome = self
-            .do_terminate_via_cap(&args.engine_id, args.grace_ms, None)
-            .await?;
-
-        let result = TerminateResult {
-            engine_id: args.engine_id,
-            sigkilled: outcome.sigkilled,
-            exit_code: outcome.exit_code,
-        };
-        serde_json::to_string(&result).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    /// Mail `aether.process.terminate` to the cap and await its
-    /// reply. Shared by the public [`Hub::terminate_substrate`] tool
-    /// and the spawn-substrate cleanup path that fires when a
-    /// component preload fails. Returns the wire `Ok` payload on
-    /// success.
-    async fn do_terminate_via_cap(
-        &self,
-        engine_id: &str,
-        grace_ms: Option<u32>,
-        timeout_ms: Option<u32>,
-    ) -> Result<TerminateOutcomeWire, McpError> {
-        let params = serde_json::json!({
-            "engine_id": engine_id,
-            "grace_ms": grace_ms,
-        });
-
-        let (_guard, rx) = self
-            .session
-            .replies
-            .register(WireTerminateResult::NAME.to_owned())
-            .ok_or_else(|| {
-                McpError::invalid_params(
-                    "a terminate_substrate is already in flight on this session; \
-                     wait for it to complete"
-                        .to_owned(),
-                    None,
-                )
-            })?;
-
-        let spec = MailSpec {
-            engine_id: HUB_SELF_ENGINE_ID.0.to_string(),
-            recipient_name: ProcessCapability::NAMESPACE.to_owned(),
-            kind_name: WireTerminate::NAME.to_owned(),
-            params: Some(params),
-            count: 1,
-        };
-        deliver_one(spec, &self.state.engines, self.session.token)
-            .await
-            .map_err(|e| McpError::internal_error(format!("send failed: {e}"), None))?;
-
-        let wait = timeout_ms
-            .map(|ms| Duration::from_millis(ms as u64).min(MAX_REPLY_TIMEOUT))
-            .unwrap_or(DEFAULT_REPLY_TIMEOUT);
-        let queued = match timeout(wait, rx).await {
-            Ok(Ok(m)) => m,
-            Ok(Err(_)) => {
-                return Err(McpError::internal_error(
-                    "terminate_result channel closed before reply arrived".to_owned(),
-                    None,
-                ));
-            }
-            Err(_) => {
-                return Err(McpError::internal_error(
-                    format!(
-                        "timed out after {}ms waiting for terminate_result",
-                        wait.as_millis()
-                    ),
-                    None,
-                ));
-            }
-        };
-
-        match postcard::from_bytes::<WireTerminateResult>(&queued.payload)
-            .map_err(|e| McpError::internal_error(format!("decode reply: {e}"), None))?
-        {
-            WireTerminateResult::Ok {
-                exit_code,
-                sigkilled,
-            } => Ok(TerminateOutcomeWire {
-                exit_code,
-                sigkilled,
-            }),
-            WireTerminateResult::Err { error } => Err(McpError::internal_error(
-                format!("terminate failed: {error}"),
-                None,
-            )),
-        }
-    }
-
-    #[tool(
-        description = "Drain captured substrate log entries for an engine (ADR-0023). Cursor-based polling: pass back `next_since` from the previous response to receive only new entries. `level` (default `\"trace\"`) filters server-side. `max` defaults to 100, clamped to 1000. `truncated_before` is set when the hub-side ring evicted entries the caller hadn't seen — treat it as a signal to poll more often or accept the gap. Buffer survives engine exit until hub shutdown, so post-mortem polls work after the substrate has crashed."
-    )]
-    pub(super) async fn engine_logs(
-        &self,
-        Parameters(args): Parameters<EngineLogsArgs>,
-    ) -> Result<String, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        let min_level = match args.level.as_deref() {
-            None | Some("trace") => LogLevel::Trace,
-            Some("debug") => LogLevel::Debug,
-            Some("info") => LogLevel::Info,
-            Some("warn") => LogLevel::Warn,
-            Some("error") => LogLevel::Error,
-            Some(other) => {
-                return Err(McpError::invalid_params(
-                    format!("level {other:?} is not one of trace|debug|info|warn|error",),
-                    None,
-                ));
-            }
-        };
-        let max = args
-            .max
-            .map(|n| n as usize)
-            .unwrap_or(TOOL_DEFAULT_ENTRIES)
-            .min(TOOL_MAX_ENTRIES);
-        let since = args.since.unwrap_or(0);
-        let result = self.state.logs.read(id, max, min_level, since);
-        let response = EngineLogsResponse {
-            engine_id: args.engine_id,
-            entries: result
-                .entries
-                .into_iter()
-                .map(|e| EngineLogEntry {
-                    timestamp_unix_ms: e.timestamp_unix_ms,
-                    level: log_level_to_str(e.level).to_owned(),
-                    target: e.target,
-                    message: e.message,
-                    sequence: e.sequence,
-                })
-                .collect(),
-            next_since: result.next_since,
-            truncated_before: result.truncated_before,
-        };
-        serde_json::to_string(&response).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    #[tool(
-        description = "Capture the current frame contents of the given engine as a PNG and return it inline as MCP image content. The substrate renders to an offscreen target so the capture works regardless of window visibility (important on macOS where occluded windows would otherwise block). Optionally carries two mail bundles dispatched atomically around the capture: `mails` fires *before* the frame is read back (state-changing mail whose effects should appear in the image), and `after_mails` fires *after* readback (cleanup such as restoring a flag the caller flipped for the capture). The substrate resolves every envelope in both bundles first (abort-on-first-failure); any failure means no mail moves. The render loop's mail drain barrier ensures the pre-capture bundle has been fully processed before the frame is read back. Empty bundles mean \"just capture\" / \"no cleanup\". Rejects if another capture is already in flight on this session, or if the wait exceeds `timeout_ms` (default 5000, clamped to 30000). Use this to verify substrate rendering end-to-end and to bundle \"set X, show me, restore X\" into one atomic tool call."
-    )]
-    pub(super) async fn capture_frame(
-        &self,
-        Parameters(args): Parameters<CaptureFrameArgs>,
-    ) -> Result<CallToolResult, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        let record = self.state.engines.get(&id).ok_or_else(|| {
-            McpError::invalid_params(format!("unknown engine_id {}", args.engine_id), None)
-        })?;
-
-        // Encode each envelope in both bundles against its kind's
-        // descriptor. Done before we register the reply-waiter or
-        // send anything, so a bad bundle produces a clean invalid-
-        // params error and never touches the engine wire.
-        let envelopes = encode_capture_bundle(&args.mails, &record).map_err(|e| {
-            McpError::invalid_params(format!("capture_frame mails bundle: {e}"), None)
-        })?;
-        let after_envelopes = encode_capture_bundle(&args.after_mails, &record).map_err(|e| {
-            McpError::invalid_params(format!("capture_frame after_mails bundle: {e}"), None)
-        })?;
-        let bundle_params = serde_json::json!({
-            "mails": envelopes,
-            "after_mails": after_envelopes,
-        });
-
-        // Register the reply-waiter BEFORE sending the request, so the
-        // engine reader can divert the reply to our oneshot the moment
-        // it arrives. If another capture is already pending on this
-        // session, surface it as a clear error rather than silently
-        // waiting behind the first.
-        let (_guard, rx) = self
-            .session
-            .replies
-            .register(KIND_CAPTURE_FRAME_RESULT.to_owned())
-            .ok_or_else(|| {
-                McpError::invalid_params(
-                    "a capture is already in flight on this session; wait for it to complete"
-                        .to_owned(),
-                    None,
-                )
-            })?;
-
-        // Send the capture_frame request carrying the pre-encoded
-        // bundle. The substrate's schema decoder reconstructs each
-        // envelope's `payload: Vec<u8>` from the JSON byte array the
-        // hub encoder wrote — postcard-equivalent round-trip.
-        let spec = MailSpec {
-            engine_id: args.engine_id.clone(),
-            recipient_name: "aether.render".to_owned(),
-            kind_name: KIND_CAPTURE_FRAME.to_owned(),
-            params: Some(bundle_params),
-            count: 1,
-        };
-        deliver_one(spec, &self.state.engines, self.session.token)
-            .await
-            .map_err(|e| McpError::internal_error(format!("send failed: {e}"), None))?;
-
-        // Wait for the reply (or timeout). `_guard` drops when this
-        // future resolves — on any path — which removes the registry
-        // entry. The record lookup above already pinned `record` for
-        // the duration of the send; we don't need it past this point.
-        let _ = record;
-        let wait = args
-            .timeout_ms
-            .map(|ms| Duration::from_millis(ms as u64).min(MAX_CAPTURE_TIMEOUT))
-            .unwrap_or(DEFAULT_CAPTURE_TIMEOUT);
-        let queued = match timeout(wait, rx).await {
-            Ok(Ok(m)) => m,
-            Ok(Err(_)) => {
-                return Err(McpError::internal_error(
-                    "capture reply channel closed before reply arrived".to_owned(),
-                    None,
-                ));
-            }
-            Err(_) => {
-                return Err(McpError::internal_error(
-                    format!(
-                        "timed out after {}ms waiting for capture_frame_result",
-                        wait.as_millis()
-                    ),
-                    None,
-                ));
-            }
-        };
-
-        // Decode the reply. The payload is a postcard-encoded
-        // `CaptureFrameResult` — either `Ok { png }` or `Err { error }`.
-        let result: CaptureFrameResultWire = postcard::from_bytes(&queued.payload)
-            .map_err(|e| McpError::internal_error(format!("decode reply: {e}"), None))?;
-        match result {
-            CaptureFrameResultWire::Err { error } => Err(McpError::internal_error(
-                format!("substrate capture failed: {error}"),
-                None,
-            )),
-            CaptureFrameResultWire::Ok { png } => {
-                let encoded = base64::engine::general_purpose::STANDARD.encode(&png);
-                Ok(CallToolResult::success(vec![Content::image(
-                    encoded,
-                    "image/png",
-                )]))
-            }
-        }
-    }
-
-    #[tool(
-        description = "Load a WASM component into the substrate by filesystem path. The hub reads the binary from `binary_path`, forwards the bytes to the substrate as `aether.component.load`, and waits for the substrate's `LoadResult` reply — returning `{mailbox_id, name}` inline or an error. Path must exist as given (no `~` expansion, no relative resolution — same rule as `spawn_substrate`). The component's kind vocabulary rides in the wasm's `aether.kinds` custom section (ADR-0028) — the loader doesn't declare kinds separately. Rejects with \"already in flight\" if a prior `load_component` on this session hasn't completed. Default timeout 5000ms; clamped to 30000. Agents never inline the wasm bytes through the tool call — that's what the path is for."
-    )]
-    pub(super) async fn load_component(
-        &self,
-        Parameters(args): Parameters<LoadComponentArgs>,
-    ) -> Result<String, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        if self.state.engines.get(&id).is_none() {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        }
-
-        let response = self
-            .do_load_component(
-                &id,
-                &args.engine_id,
-                &args.binary_path,
-                args.name,
-                args.timeout_ms,
-            )
-            .await?;
-        serde_json::to_string(&response).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    /// Shared load-component core. `load_component` wraps this after
-    /// parsing the engine id from the MCP arg; `spawn_substrate` calls
-    /// it directly once per entry in its `components` preload list.
-    /// Reads the wasm file, dispatches `aether.component.load`
-    /// against the usual reply-guard, decodes the response, and
-    /// updates the engine registry on success — returning the same
-    /// `LoadComponentResponse` shape the public tool returns.
-    async fn do_load_component(
-        &self,
-        engine_id: &EngineId,
-        engine_id_str: &str,
-        binary_path: &str,
-        name: Option<String>,
-        timeout_ms: Option<u32>,
-    ) -> Result<LoadComponentResponse, McpError> {
-        let wasm = tokio::fs::read(binary_path).await.map_err(|e| {
-            McpError::invalid_params(format!("reading binary_path {binary_path:?}: {e}"), None)
-        })?;
-
-        let params = serde_json::json!({
-            "wasm": wasm,
-            "name": name,
-        });
-
-        let (_guard, rx) = self
-            .session
-            .replies
-            .register(KIND_LOAD_RESULT.to_owned())
-            .ok_or_else(|| {
-                McpError::invalid_params(
-                    "a load_component is already in flight on this session; wait for it to complete"
-                        .to_owned(),
-                    None,
-                )
-            })?;
-
-        let spec = MailSpec {
-            engine_id: engine_id_str.to_owned(),
-            recipient_name: "aether.component".to_owned(),
-            kind_name: KIND_LOAD_COMPONENT.to_owned(),
-            params: Some(params),
-            count: 1,
-        };
-        deliver_one(spec, &self.state.engines, self.session.token)
-            .await
-            .map_err(|e| McpError::internal_error(format!("send failed: {e}"), None))?;
-
-        let wait = timeout_ms
-            .map(|ms| Duration::from_millis(ms as u64).min(MAX_REPLY_TIMEOUT))
-            .unwrap_or(DEFAULT_REPLY_TIMEOUT);
-        let queued = match timeout(wait, rx).await {
-            Ok(Ok(m)) => m,
-            Ok(Err(_)) => {
-                return Err(McpError::internal_error(
-                    "load_result channel closed before reply arrived".to_owned(),
-                    None,
-                ));
-            }
-            Err(_) => {
-                return Err(McpError::internal_error(
-                    format!(
-                        "timed out after {}ms waiting for load_result",
-                        wait.as_millis()
-                    ),
-                    None,
-                ));
-            }
-        };
-
-        let result: LoadResultWire = postcard::from_bytes(&queued.payload)
-            .map_err(|e| McpError::internal_error(format!("decode reply: {e}"), None))?;
-        match result {
-            LoadResultWire::Err { error } => Err(McpError::internal_error(
-                format!("substrate load failed: {error}"),
-                None,
-            )),
-            LoadResultWire::Ok {
-                mailbox_id,
-                name,
-                capabilities,
-            } => {
-                self.state.engines.upsert_component(
-                    engine_id,
-                    aether_data::MailboxId(mailbox_id),
-                    ComponentRecord {
-                        name: name.clone(),
-                        capabilities: capabilities.clone(),
-                    },
-                );
-                // ADR-0064: encode the raw u64 the substrate returned
-                // into the tagged-string form so the agent never sees
-                // the precision-losing JSON number on the wire.
-                let mailbox_id = tagged_id::encode(mailbox_id).ok_or_else(|| {
-                    McpError::internal_error(
-                        format!("substrate returned untagged mailbox id {mailbox_id:#x}"),
-                        None,
-                    )
-                })?;
-                Ok(LoadComponentResponse {
-                    mailbox_id,
-                    name,
-                    capabilities,
-                })
-            }
-        }
-    }
-
-    #[tool(
-        description = "Atomically replace a live component's WASM with a new binary loaded from a filesystem path (ADR-0022 + ADR-0038 structural splice). The hub reads the binary from `binary_path` and forwards `aether.component.replace` to the substrate, which splices the mailbox's inbox over to a new dispatcher, drops the old sender, and joins the old dispatcher once it drains naturally. Kind vocabulary rides in the wasm's `aether.kinds` custom section (ADR-0028). `drain_timeout_ms` is accepted for wire compatibility but currently ignored — post-ADR-0038 there is no freeze-drain timeout to enforce. Path must exist as given. Waits for the substrate's `ReplaceResult`; returns `\"Ok\"` on success. Rejects with \"already in flight\" if a prior replace is pending on this session. Tool timeout default 5000ms, clamped to 30000."
-    )]
-    pub(super) async fn replace_component(
-        &self,
-        Parameters(args): Parameters<ReplaceComponentArgs>,
-    ) -> Result<String, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        if self.state.engines.get(&id).is_none() {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        }
-        // ADR-0064: decode the tagged-string mailbox id back to u64
-        // for the substrate-side wire (the substrate keeps internal
-        // ids raw; the prefix-string form is purely the MCP boundary).
-        let mailbox_id = tagged_id::decode_with_tag(&args.mailbox_id, Tag::Mailbox)
-            .map_err(|e| McpError::invalid_params(format!("mailbox_id: {e}"), None))?;
-
-        let wasm = tokio::fs::read(&args.binary_path).await.map_err(|e| {
-            McpError::invalid_params(
-                format!("reading binary_path {:?}: {e}", args.binary_path),
-                None,
-            )
-        })?;
-
-        let params = serde_json::json!({
-            "mailbox_id": mailbox_id,
-            "wasm": wasm,
-            "drain_timeout_ms": args.drain_timeout_ms,
-        });
-
-        let (_guard, rx) = self
-            .session
-            .replies
-            .register(KIND_REPLACE_RESULT.to_owned())
-            .ok_or_else(|| {
-                McpError::invalid_params(
-                    "a replace_component is already in flight on this session; wait for it to complete"
-                        .to_owned(),
-                    None,
-                )
-            })?;
-
-        let spec = MailSpec {
-            engine_id: args.engine_id.clone(),
-            recipient_name: "aether.component".to_owned(),
-            kind_name: KIND_REPLACE_COMPONENT.to_owned(),
-            params: Some(params),
-            count: 1,
-        };
-        deliver_one(spec, &self.state.engines, self.session.token)
-            .await
-            .map_err(|e| McpError::internal_error(format!("send failed: {e}"), None))?;
-
-        let wait = args
-            .timeout_ms
-            .map(|ms| Duration::from_millis(ms as u64).min(MAX_REPLY_TIMEOUT))
-            .unwrap_or(DEFAULT_REPLY_TIMEOUT);
-        let queued = match timeout(wait, rx).await {
-            Ok(Ok(m)) => m,
-            Ok(Err(_)) => {
-                return Err(McpError::internal_error(
-                    "replace_result channel closed before reply arrived".to_owned(),
-                    None,
-                ));
-            }
-            Err(_) => {
-                return Err(McpError::internal_error(
-                    format!(
-                        "timed out after {}ms waiting for replace_result",
-                        wait.as_millis()
-                    ),
-                    None,
-                ));
-            }
-        };
-
-        let result: ReplaceResultWire = postcard::from_bytes(&queued.payload)
-            .map_err(|e| McpError::internal_error(format!("decode reply: {e}"), None))?;
-        match result {
-            ReplaceResultWire::Err { error } => Err(McpError::internal_error(
-                format!("substrate replace failed: {error}"),
-                None,
-            )),
-            ReplaceResultWire::Ok { capabilities } => {
-                // ADR-0033: the replaced component may advertise a
-                // different receive surface. Refresh the cached record
-                // so `describe_component` reflects what's actually
-                // bound now. Name is preserved — `replace_component`
-                // keeps the existing mailbox + name by design.
-                let mailbox = aether_data::MailboxId(mailbox_id);
-                let existing = self.state.engines.get_component(&id, mailbox);
-                let name = existing
-                    .map(|r| r.name)
-                    .unwrap_or_else(|| format!("mailbox_{}", args.mailbox_id));
-                self.state.engines.upsert_component(
-                    &id,
-                    mailbox,
-                    ComponentRecord {
-                        name,
-                        capabilities: capabilities.clone(),
-                    },
-                );
-                serde_json::to_string(&capabilities)
-                    .map_err(|e| McpError::internal_error(e.to_string(), None))
-            }
-        }
-    }
-
-    #[tool(
-        description = "Describe a loaded component's receive-side capabilities (ADR-0033). Returns the component's name, its top-level author-written documentation, the set of kinds it typed-handles (id, name, optional per-handler doc), and whether it has a `#[fallback]` catchall (with the fallback's own optional doc). The capability set is parsed from the component's `aether.kinds.inputs` wasm custom section at `load_component` / `replace_component` time. Strict receivers — components without a fallback — are distinguishable via `fallback: null` in the response. Components predating the `#[actor]` macro ship with empty fields since they have no structured capability surface to advertise."
-    )]
-    pub(super) async fn describe_component(
-        &self,
-        Parameters(args): Parameters<DescribeComponentArgs>,
-    ) -> Result<String, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        if self.state.engines.get(&id).is_none() {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        }
-        // ADR-0064: agents pass the tagged-string form (`mbx-XXXX-...`).
-        // Decode here so the registry lookup uses the raw u64.
-        let mailbox_id = tagged_id::decode_with_tag(&args.mailbox_id, Tag::Mailbox)
-            .map_err(|e| McpError::invalid_params(format!("mailbox_id: {e}"), None))?;
-        let Some(component) = self
-            .state
-            .engines
-            .get_component(&id, aether_data::MailboxId(mailbox_id))
-        else {
-            return Err(McpError::invalid_params(
-                format!(
-                    "no component at mailbox_id {} on engine {}",
-                    args.mailbox_id, args.engine_id
-                ),
-                None,
-            ));
-        };
-        let ComponentRecord { name, capabilities } = component;
-        let response = DescribeComponentResponse {
-            name,
-            doc: capabilities.doc,
-            receives: capabilities.handlers,
-            fallback: capabilities.fallback,
-        };
-        serde_json::to_string(&response).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    #[tool(
-        description = "Export the mail subtree under one root as a trace JSON file (issue iamacoffeepot/aether#728, ADR-0080 Phase 3). Output is the Chrome trace event format, which Perfetto, chrome://tracing, and speedscope all read natively — drop the file into perfetto.dev for a flame graph + causal arrows view. Calls the same `describe_tree` substrate path under the hood, then re-shapes each mail into one `ph:\"X\"` (complete) event covering its `t_received → t_finished` interval on the recipient's lane, plus `ph:\"s\"`/`ph:\"f\"` (flow) arrows linking parent.t_finished → child.t_received for causal visualization. Mails without `t_received`/`t_finished` (orphan or in-flight at query time) are skipped. When `output_path` is omitted, returns the JSON inline (fine for small subtrees, may overflow agent context for busy substrates). When `output_path` is set, writes to the path (creating parent dirs) and returns a `{path}` confirmation. Default timeout 5000ms; clamped to 30000."
-    )]
-    pub(super) async fn dump_trace(
-        &self,
-        Parameters(args): Parameters<DumpTraceArgs>,
-    ) -> Result<String, McpError> {
-        // Build kind-id → name + mailbox-id → (name, category) lookups
-        // from the engine's handshake-time + post-load caches. The
-        // renderer uses the human-readable forms in event `name` /
-        // id fields; missing entries fall back to the tagged-string
-        // id so an out-of-sync inventory surfaces as visible drift
-        // rather than silent corruption (issue
-        // iamacoffeepot/aether#731).
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        let Some(record) = self.state.engines.get(&id) else {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        };
-        let kind_names: std::collections::HashMap<u64, String> = record
-            .kinds
-            .iter()
-            .map(|k| {
-                let kid = aether_data::canonical::kind_id_from_parts(&k.name, &k.schema);
-                (kid, k.name.clone())
-            })
-            .collect();
-        let mailbox_names: super::trace::MailboxLookup = record
-            .mailboxes
-            .iter()
-            .map(|m| (m.id.0, (m.name.clone(), m.category)))
-            .collect();
-        drop(record);
-
-        let describe_args = DescribeTreeArgs {
-            engine_id: args.engine_id.clone(),
-            root: args.root,
-            timeout_ms: args.timeout_ms,
-        };
-        let result = self.do_describe_tree(describe_args).await?;
-        let trace_json = render(&result, &kind_names, &mailbox_names)
-            .map_err(|e| McpError::internal_error(format!("render trace: {e}"), None))?;
-
-        match args.output_path {
-            Some(path) => {
-                if let Some(parent) = std::path::Path::new(&path).parent()
-                    && !parent.as_os_str().is_empty()
-                {
-                    tokio::fs::create_dir_all(parent).await.map_err(|e| {
-                        McpError::internal_error(
-                            format!("create parent dirs for {path}: {e}"),
-                            None,
-                        )
-                    })?;
-                }
-                tokio::fs::write(&path, &trace_json).await.map_err(|e| {
-                    McpError::internal_error(format!("write trace to {path}: {e}"), None)
-                })?;
-                Ok(serde_json::json!({ "path": path }).to_string())
-            }
-            None => Ok(trace_json),
-        }
-    }
-
-    #[tool(
-        description = "Describe the mail tree under one root (issue 718, ADR-0080 Phase 2). The substrate's `aether.trace` observer accumulates a parent → child mail graph keyed by `MailId`; this tool surfaces one root's subtree as JSON: every node's `mail_id`, `parent`, `sender`, `recipient`, `kind`, and `t_sent` / `t_received` / `t_finished` timestamps in nanoseconds-since-substrate-boot, plus the root's current `in_flight` count. `root` is a JSON object `{ sender: \"mbx-XXXX-XXXX-XXXX\", correlation_id: <number> }` — copy a `MailNodeWire.root` from a prior `describe_tree` reply or a `RootSummaryWire.root` from `list_active_roots`. Returns `Err::not_found` (echoing the requested root) when the root has been evicted past retention or never existed. Default timeout 5000ms; clamped to 30000."
-    )]
-    pub(super) async fn describe_tree(
-        &self,
-        Parameters(args): Parameters<DescribeTreeArgs>,
-    ) -> Result<String, McpError> {
-        let result = self.do_describe_tree(args).await?;
-        // The cap's reply is already MailId/MailboxId/KindId/Nanos —
-        // their human-readable serde branches emit tagged strings for
-        // ids and a u64 for `Nanos`, so JSON serialization is a single
-        // shot with no manual conversion.
-        serde_json::to_string(&result).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    /// Shared describe_tree core. The public `describe_tree` tool wraps
-    /// this with JSON serialization; `dump_trace` reuses it to fetch
-    /// the same observer state and re-shape it as trace events (issue
-    /// iamacoffeepot/aether#728).
-    async fn do_describe_tree(
-        &self,
-        args: DescribeTreeArgs,
-    ) -> Result<DescribeTreeResult, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        if self.state.engines.get(&id).is_none() {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        }
-        // Validate the tagged-string sender up-front so a malformed
-        // form fails fast with a clear error rather than surfacing as
-        // a downstream encode error from the schema codec.
-        tagged_id::decode_with_tag(&args.root.sender, Tag::Mailbox)
-            .map_err(|e| McpError::invalid_params(format!("root.sender: {e}"), None))?;
-        let root_params = serde_json::json!({
-            "root": {
-                "sender": args.root.sender.clone(),
-                "correlation_id": args.root.correlation_id,
-            },
-        });
-
-        let result_kind = <DescribeTreeResult as Kind>::NAME.to_owned();
-        let (_guard, rx) = self
-            .session
-            .replies
-            .register(result_kind.clone())
-            .ok_or_else(|| {
-                McpError::invalid_params(
-                    "a describe_tree is already in flight on this session; wait for it to complete"
-                        .to_owned(),
-                    None,
-                )
-            })?;
-
-        let spec = MailSpec {
-            engine_id: args.engine_id.clone(),
-            recipient_name: TRACE_OBSERVER_MAILBOX_NAME.to_owned(),
-            kind_name: <DescribeTree as Kind>::NAME.to_owned(),
-            params: Some(root_params),
-            count: 1,
-        };
-        deliver_one(spec, &self.state.engines, self.session.token)
-            .await
-            .map_err(|e| McpError::internal_error(format!("send failed: {e}"), None))?;
-
-        let wait = args
-            .timeout_ms
-            .map(|ms| Duration::from_millis(ms as u64).min(MAX_REPLY_TIMEOUT))
-            .unwrap_or(DEFAULT_REPLY_TIMEOUT);
-        let queued = match timeout(wait, rx).await {
-            Ok(Ok(m)) => m,
-            Ok(Err(_)) => {
-                return Err(McpError::internal_error(
-                    "describe_tree reply channel closed before reply arrived".to_owned(),
-                    None,
-                ));
-            }
-            Err(_) => {
-                return Err(McpError::internal_error(
-                    format!(
-                        "timed out after {}ms waiting for {result_kind}",
-                        wait.as_millis()
-                    ),
-                    None,
-                ));
-            }
-        };
-
-        postcard::from_bytes(&queued.payload)
-            .map_err(|e| McpError::internal_error(format!("decode reply: {e}"), None))
-    }
-
-    #[tool(
-        description = "Export every mail whose `t_sent` falls within a time window as a trace JSON file (issue iamacoffeepot/aether#735, ADR-0080 Phase 3). Same Chrome trace event format as `dump_trace`, just collected by overlap rather than by root walk — useful for \"what happened in the last N ms?\" perf investigation. Loads into Perfetto, chrome://tracing, or speedscope. `window` is either `{\"last_ms\": N}` for last-N-milliseconds (resolved against substrate monotonic now) or `{\"start_ns\": ..., \"end_ns\": ...}` for an explicit absolute range. Strict `t_sent` containment — parent edges may dangle to mail outside the window (Perfetto renders these as flow arrows with no destination); drill into a specific root via `dump_trace` for full chain context. `max_mails` caps the substrate's matched set (default 10_000, hard cap 100_000); over-cap windows fail with the matched count so the caller can narrow before retrying. When `output_path` is omitted the JSON is returned inline (likely to overflow agent context for any non-trivial window — point at a file). When set, writes to the path (creating parent dirs) and returns a `{path}` confirmation. Default timeout 5000ms; clamped to 30000."
-    )]
-    pub(super) async fn dump_trace_window(
-        &self,
-        Parameters(args): Parameters<DumpTraceWindowArgs>,
-    ) -> Result<String, McpError> {
-        // Same lookup-build path as `dump_trace`. Resolved entries
-        // give human-readable kind names + category-prefixed mailbox
-        // names in the trace doc; missing entries fall back to the
-        // tagged-string id so an out-of-sync inventory surfaces as
-        // visible drift rather than silent corruption (issue 731).
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        let Some(record) = self.state.engines.get(&id) else {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        };
-        let kind_names: std::collections::HashMap<u64, String> = record
-            .kinds
-            .iter()
-            .map(|k| {
-                let kid = aether_data::canonical::kind_id_from_parts(&k.name, &k.schema);
-                (kid, k.name.clone())
-            })
-            .collect();
-        let mailbox_names: super::trace::MailboxLookup = record
-            .mailboxes
-            .iter()
-            .map(|m| (m.id.0, (m.name.clone(), m.category)))
-            .collect();
-        drop(record);
-
-        let describe_args = DescribeTreeWindowArgs {
-            engine_id: args.engine_id.clone(),
-            window: args.window,
-            max_mails: args.max_mails,
-            timeout_ms: args.timeout_ms,
-        };
-        let result = self.do_describe_window(describe_args).await?;
-        let mails = match result {
-            DescribeWindowResult::Ok { mails } => mails,
-            DescribeWindowResult::Err { too_many } => {
-                let count = too_many
-                    .map(|n| n.to_string())
-                    .unwrap_or_else(|| "<unspecified>".to_owned());
-                return Err(McpError::invalid_params(
-                    format!(
-                        "describe_window matched {count} mails (over max_mails); narrow the window or raise max_mails"
-                    ),
-                    None,
-                ));
-            }
-        };
-        let trace_json = render_mails(&mails, &kind_names, &mailbox_names)
-            .map_err(|e| McpError::internal_error(format!("render trace: {e}"), None))?;
-
-        match args.output_path {
-            Some(path) => {
-                if let Some(parent) = std::path::Path::new(&path).parent()
-                    && !parent.as_os_str().is_empty()
-                {
-                    tokio::fs::create_dir_all(parent).await.map_err(|e| {
-                        McpError::internal_error(
-                            format!("create parent dirs for {path}: {e}"),
-                            None,
-                        )
-                    })?;
-                }
-                tokio::fs::write(&path, &trace_json).await.map_err(|e| {
-                    McpError::internal_error(format!("write trace to {path}: {e}"), None)
-                })?;
-                Ok(serde_json::json!({ "path": path }).to_string())
-            }
-            None => Ok(trace_json),
-        }
-    }
-
-    #[tool(
-        description = "Describe every mail whose `t_sent` falls within a time window (issue iamacoffeepot/aether#735). Same observer state as `describe_tree`, just collected by overlap rather than by root walk — useful for perf investigation when you want \"what happened in the last N ms?\" rather than drilling into one specific chain. `window` is either `{\"last_ms\": N}` for last-N-milliseconds (resolved against the substrate's monotonic now) or `{\"start_ns\": ..., \"end_ns\": ...}` for an explicit absolute range (`end_ns` optional, defaults to \"now\"). Strict `t_sent` containment — parent edges may dangle to mail outside the window; drill into a specific root via `describe_tree` for full chain context. `max_mails` caps the reply (default 10_000, substrate hard cap 100_000); over-cap windows reply `{\"Err\": {\"too_many\": <matched_count>}}` instead of truncating, so the caller can narrow before retrying. Default timeout 5000ms; clamped to 30000."
-    )]
-    pub(super) async fn describe_tree_window(
-        &self,
-        Parameters(args): Parameters<DescribeTreeWindowArgs>,
-    ) -> Result<String, McpError> {
-        let result = self.do_describe_window(args).await?;
-        // `DescribeWindowResult`'s serde branches emit tagged strings
-        // for ids and a u64 for `Nanos` on human-readable serializers,
-        // so JSON serialization is a single shot. The externally-
-        // tagged `Ok` / `Err` shape mirrors `describe_tree`'s reply.
-        serde_json::to_string(&result).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    /// Shared describe_tree_window core. The public tool wraps this
-    /// with JSON serialization; `dump_trace_window` (Phase 3) reuses
-    /// it to fetch the same observer state and re-shape it as trace
-    /// events.
-    pub(super) async fn do_describe_window(
-        &self,
-        args: DescribeTreeWindowArgs,
-    ) -> Result<DescribeWindowResult, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        if self.state.engines.get(&id).is_none() {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        }
-
-        // Convert the agent's untagged `TraceWindowArg` to the
-        // substrate-side externally-tagged `TraceWindow`, then
-        // serialize the whole `DescribeWindow` request through serde.
-        // The schema codec expects externally-tagged enum JSON
-        // (`{"Absolute": {...}}` / `{"Relative": {...}}`), which is
-        // exactly what serde's default impl emits for `TraceWindow`.
-        let window_kind = match args.window {
-            TraceWindowArg::Absolute { start_ns, end_ns } => {
-                TraceWindow::Absolute { start_ns, end_ns }
-            }
-            TraceWindowArg::Relative { last_ms } => TraceWindow::Relative { last_ms },
-        };
-        let request = DescribeWindow {
-            window: window_kind,
-            max_mails: args.max_mails,
-        };
-        let request_params = serde_json::to_value(request).map_err(|e| {
-            McpError::internal_error(format!("encode describe_window request: {e}"), None)
-        })?;
-
-        let result_kind = <DescribeWindowResult as Kind>::NAME.to_owned();
-        let (_guard, rx) = self
-            .session
-            .replies
-            .register(result_kind.clone())
-            .ok_or_else(|| {
-                McpError::invalid_params(
-                    "a describe_tree_window is already in flight on this session; wait for it to complete"
-                        .to_owned(),
-                    None,
-                )
-            })?;
-
-        let spec = MailSpec {
-            engine_id: args.engine_id.clone(),
-            recipient_name: TRACE_OBSERVER_MAILBOX_NAME.to_owned(),
-            kind_name: <DescribeWindow as Kind>::NAME.to_owned(),
-            params: Some(request_params),
-            count: 1,
-        };
-        deliver_one(spec, &self.state.engines, self.session.token)
-            .await
-            .map_err(|e| McpError::internal_error(format!("send failed: {e}"), None))?;
-
-        let wait = args
-            .timeout_ms
-            .map(|ms| Duration::from_millis(ms as u64).min(MAX_REPLY_TIMEOUT))
-            .unwrap_or(DEFAULT_REPLY_TIMEOUT);
-        let queued = match timeout(wait, rx).await {
-            Ok(Ok(m)) => m,
-            Ok(Err(_)) => {
-                return Err(McpError::internal_error(
-                    "describe_tree_window reply channel closed before reply arrived".to_owned(),
-                    None,
-                ));
-            }
-            Err(_) => {
-                return Err(McpError::internal_error(
-                    format!(
-                        "timed out after {}ms waiting for {result_kind}",
-                        wait.as_millis()
-                    ),
-                    None,
-                ));
-            }
-        };
-
-        postcard::from_bytes(&queued.payload)
-            .map_err(|e| McpError::internal_error(format!("decode reply: {e}"), None))
-    }
-
-    #[tool(
-        description = "List recent root mails the substrate's `aether.trace` observer is tracking (issue 718, ADR-0080 Phase 2). Surfaces one summary per active root: `root` (composite `{ sender, correlation_id }`), `kind` (tagged-string id of the root mail's payload kind), `sender` and `recipient` (tagged-string mailbox ids), `t_sent` (nanoseconds since substrate boot), and `in_flight` (count of mails in this chain currently between `Sent` and `Finished`). Sorted desc by `t_sent` so the most recent roots come first. `since_ms` filters by the root's originating send time (default 60_000); `max` caps the reply length (default 50, hard cap 1000). Use this to discover root ids the agent can hand to `describe_tree` for a full subtree view, or to spot-check whether a chain has settled by watching `in_flight` drop to zero. Default timeout 5000ms; clamped to 30000."
-    )]
-    pub(super) async fn list_active_roots(
-        &self,
-        Parameters(args): Parameters<ListActiveRootsArgs>,
-    ) -> Result<String, McpError> {
-        let uuid = Uuid::parse_str(&args.engine_id).map_err(|e| {
-            McpError::invalid_params(format!("engine_id is not a valid UUID: {e}"), None)
-        })?;
-        let id = EngineId(uuid);
-        if self.state.engines.get(&id).is_none() {
-            return Err(McpError::invalid_params(
-                format!("unknown engine_id {}", args.engine_id),
-                None,
-            ));
-        }
-
-        // ListActiveRoots's two fields are both `Option<u32>` — the
-        // schema-driven encoder accepts JSON `null` for None and the
-        // raw number for Some, so we serialize the agent's args
-        // straight through.
-        let request_params = serde_json::json!({
-            "since_ms": args.since_ms,
-            "max": args.max,
-        });
-
-        let result_kind = <ListActiveRootsResult as Kind>::NAME.to_owned();
-        let (_guard, rx) = self
-            .session
-            .replies
-            .register(result_kind.clone())
-            .ok_or_else(|| {
-                McpError::invalid_params(
-                    "a list_active_roots is already in flight on this session; wait for it to complete"
-                        .to_owned(),
-                    None,
-                )
-            })?;
-
-        let spec = MailSpec {
-            engine_id: args.engine_id.clone(),
-            recipient_name: TRACE_OBSERVER_MAILBOX_NAME.to_owned(),
-            kind_name: <ListActiveRoots as Kind>::NAME.to_owned(),
-            params: Some(request_params),
-            count: 1,
-        };
-        deliver_one(spec, &self.state.engines, self.session.token)
-            .await
-            .map_err(|e| McpError::internal_error(format!("send failed: {e}"), None))?;
-
-        let wait = args
-            .timeout_ms
-            .map(|ms| Duration::from_millis(ms as u64).min(MAX_REPLY_TIMEOUT))
-            .unwrap_or(DEFAULT_REPLY_TIMEOUT);
-        let queued = match timeout(wait, rx).await {
-            Ok(Ok(m)) => m,
-            Ok(Err(_)) => {
-                return Err(McpError::internal_error(
-                    "list_active_roots reply channel closed before reply arrived".to_owned(),
-                    None,
-                ));
-            }
-            Err(_) => {
-                return Err(McpError::internal_error(
-                    format!(
-                        "timed out after {}ms waiting for {result_kind}",
-                        wait.as_millis()
-                    ),
-                    None,
-                ));
-            }
-        };
-
-        let result: ListActiveRootsResult = postcard::from_bytes(&queued.payload)
-            .map_err(|e| McpError::internal_error(format!("decode reply: {e}"), None))?;
-        serde_json::to_string(&result).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-
-    #[tool(
-        description = "List all engines currently connected to the hub. Each item reports the hub-assigned engine_id, the engine's self-declared name/pid/version, and a `spawned` flag: `true` if the hub launched this engine as a child process (ADR-0009), `false` if it connected externally."
-    )]
-    pub(super) async fn list_engines(&self) -> Result<String, McpError> {
-        let engines: Vec<EngineInfo> = self
-            .state
-            .engines
-            .list()
-            .into_iter()
-            .map(|r| EngineInfo {
-                engine_id: r.id.0.to_string(),
-                name: r.name,
-                pid: r.pid,
-                version: r.version,
-                spawned: r.spawned,
-            })
-            .collect();
-        serde_json::to_string(&engines).map_err(|e| McpError::internal_error(e.to_string(), None))
-    }
-}
-
-#[tool_handler]
-impl ServerHandler for Hub {
-    fn get_info(&self) -> ServerInfo {
-        ServerInfo {
-            capabilities: ServerCapabilities::builder().enable_tools().build(),
-            server_info: Implementation {
-                name: "aether-hub".into(),
-                version: env!("CARGO_PKG_VERSION").into(),
-                ..Default::default()
-            },
-            ..Default::default()
-        }
-    }
-}
