+++ crates/aether-substrate-bundle/src/hub/mod.rs
@@ -27,8 +27,6 @@
 //! - [`HubChassis`] / [`HubServerDriverCapability`] — Chassis marker +
 //!   driver capability that owns the tokio runtime + listeners.
 //! - [`run_engine_listener`] — the engine-facing TCP listener loop.
-//! - [`run_mcp_server`] — the rmcp-driven MCP transport on
-//!   [`DEFAULT_MCP_PORT`].
 //! - [`EngineRegistry`] / [`SessionRegistry`] / [`PendingSpawns`] /
 //!   [`LogStore`] — the in-process state the coordinator wires
 //!   together.
@@ -65,7 +63,6 @@ mod chassis;
 mod engine;
 mod log_store;
 mod loopback;
-mod mcp;
 mod process_capability;
 mod registry;
 mod session;
@@ -78,7 +75,6 @@ pub use chassis::{HubChassis, HubEnv, HubServerDriverCapability, HubServerDriver
 pub use engine::READ_TIMEOUT;
 pub use log_store::{LogStore, ReadResult as LogReadResult};
 pub use loopback::{HUB_SELF_ENGINE_ID, LoopbackEngine, LoopbackHandle};
-pub use mcp::{DEFAULT_MCP_PORT, HubState, run_mcp_server};
 pub use process_capability::{ProcessCapability, ProcessCapabilityConfig};
 pub use registry::{EngineRecord, EngineRegistry};
 pub use session::{
